Streamly.Internal.Data.Array.Prim

Description

Documentation

data Array a

Construction

fromListN :: Prim a => Int -> [a] -> Array a

fromList :: Prim a => [a] -> Array a

fromStreamN :: (MonadIO m, Prim a) => Int -> SerialT m a -> m (Array a)

Create an Array from the first N elements of a stream. The array is allocated to size N, if the stream terminates before N elements then the array may hold less than N elements.

Internal

fromStream :: (MonadIO m, Prim a) => SerialT m a -> m (Array a)

Create an Array from a stream. This is useful when we want to create a single array from a stream of unknown size. writeN is at least twice as efficient when the size is already known.

Note that if the input stream is too large memory allocation for the array may fail. When the stream size is not known, arraysOf followed by processing of indvidual arrays in the resulting stream should be preferred.

Internal

writeN :: (MonadIO m, Prim a) => Int -> Fold m a (Array a)

writeN n folds a maximum of n elements from the input stream to an Array.

Internal

write :: (MonadIO m, Prim a) => Fold m a (Array a)

Fold the whole input to a single array.

Caution! Do not use this on infinite streams.

Internal

Elimination

toList :: Prim a => Array a -> [a]

Convert an Array into a list.

Internal

toStream :: (MonadIO m, IsStream t, Prim a) => Array a -> t m a

Convert an Array into a stream.

Internal

toStreamRev :: (MonadIO m, IsStream t, Prim a) => Array a -> t m a

Convert an Array into a stream in reverse order.

Internal

read :: (MonadIO m, Prim a) => Unfold m (Array a) a

Unfold an array into a stream.

Since: 0.7.0

unsafeRead :: (MonadIO m, Prim a) => Unfold m (Array a) a

Unfold an array into a stream, does not check the end of the array, the user is responsible for terminating the stream within the array bounds. For high performance application where the end condition can be determined by a terminating fold.

The following might not be true, not that the representation changed. Written in the hope that it may be faster than "read", however, in the case for which this was written, "read" proves to be faster even though the core generated with unsafeRead looks simpler.

Internal

Random Access

length :: Array a -> Int

null :: Array a -> Bool

null arr = length arr == 0

Internal

last :: Prim a => Array a -> Maybe a

last arr = readIndex arr (length arr - 1)

Internal

readIndex :: Prim a => Array a -> Int -> Maybe a

O(1) Lookup the element at the given index, starting from 0.

Internal

unsafeIndex :: Prim a => Array a -> Int -> a

Immutable Transformations

Folding Arrays

streamFold :: (MonadIO m, Prim a) => (SerialT m a -> m b) -> Array a -> m b

Fold an array using a stream fold operation.

Internal

fold :: forall m a b. (MonadIO m, Prim a) => Fold m a b -> Array a -> m b

Fold an array using a Fold.

Internal

Folds with Array as the container

Streaming array operations

concat :: (IsStream t, MonadIO m, Prim a) => t m (Array a) -> t m a

Convert a stream of arrays into a stream of their elements.

Same as the following but more efficient:

concat = S.concatMap A.read

Internal

compact :: (MonadIO m, Prim a) => Int -> SerialT m (Array a) -> SerialT m (Array a)

Coalesce adjacent arrays in incoming stream to form bigger arrays of a maximum specified size in bytes.

Internal