Streamly.Internal.Data.Array.Prim.Types

Description

Documentation

data Array a

Constructors

Array ByteArray# Int Int

unsafeFreeze :: (Prim a, MonadIO m) => Array a -> m (Array a)

unsafeFreezeWithShrink :: (Prim a, MonadIO m) => Array a -> Int -> m (Array a)

defaultChunkSize :: Int

Default maximum buffer size in bytes, for reading from and writing to IO devices, the value is 32KB minus GHC allocation overhead, which is a few bytes, so that the actual allocation is 32KB.

nil :: Prim a => Array a

Construction

spliceTwo :: (MonadIO m, Prim a) => Array a -> Array a -> m (Array a)

Splice two immutable arrays creating a new immutable array.

fromList :: Prim a => [a] -> Array a

fromListN :: Prim a => Int -> [a] -> Array a

fromStreamDN :: (MonadIO m, Prim a) => Int -> Stream m a -> m (Array a)

fromStreamD :: (MonadIO m, Prim a) => Stream m a -> m (Array a)

Streams of arrays

fromStreamDArraysOf :: (MonadIO m, Prim a) => Int -> Stream m a -> Stream m (Array a)

fromStreamArraysOf n stream groups the input stream into a stream of arrays of size n.

data FlattenState s a

Constructors

OuterLoop s

InnerLoop s !(Array a) !Int !Int

flattenArrays :: (MonadIO m, Prim a) => Stream m (Array a) -> Stream m a

flattenArraysRev :: (MonadIO m, Prim a) => Stream m (Array a) -> Stream m a

data SpliceState s arr1 arr2

Constructors

SpliceInitial s

SpliceBuffering s arr2

SpliceYielding arr1 (SpliceState s arr1 arr2)

SpliceFinish

packArraysChunksOf :: forall m a. (MonadIO m, Prim a) => Int -> Stream m (Array a) -> Stream m (Array a)

Coalesce adjacent arrays in incoming stream to form bigger arrays of a maximum specified size in bytes. Note that if a single array is bigger than the specified size we do not split it to fit. When we coalesce multiple arrays if the size would exceed the specified size we do not coalesce therefore the actual array size may be less than the specified chunk size.

Internal

lpackArraysChunksOf :: forall m a. (MonadIO m, Prim a) => Int -> Fold m (Array a) () -> Fold m (Array a) ()

splitOn :: MonadIO m => Word8 -> Stream m (Array Word8) -> Stream m (Array Word8)

Split a stream of arrays on a given separator byte, dropping the separator and coalescing all the arrays between two separators into a single array.

Internal

breakOn :: MonadIO m => Word8 -> Array Word8 -> m (Array Word8, Maybe (Array Word8))

Elimination

unsafeIndex :: Prim a => Array a -> Int -> a

byteLength :: forall a. Prim a => Array a -> Int

length :: Array a -> Int

foldl' :: Prim a => (b -> a -> b) -> b -> Array a -> b

Strict left-associated fold over the elements of an Array.

foldr :: Prim a => (a -> b -> b) -> b -> Array a -> b

foldr' :: Prim a => (a -> b -> b) -> b -> Array a -> b

Strict right-associated fold over the elements of an Array.

foldlM' :: (Prim a, Monad m) => (b -> a -> m b) -> b -> Array a -> m b

Strict left-associated fold over the elements of an Array.

splitAt :: Int -> [a] -> ([a], [a])

splitAt n xs returns a tuple where first element is xs prefix of length n and second element is the remainder of the list:

splitAt 6 "Hello World!" == ("Hello ","World!")
splitAt 3 [1,2,3,4,5] == ([1,2,3],[4,5])
splitAt 1 [1,2,3] == ([1],[2,3])
splitAt 3 [1,2,3] == ([1,2,3],[])
splitAt 4 [1,2,3] == ([1,2,3],[])
splitAt 0 [1,2,3] == ([],[1,2,3])
splitAt (-1) [1,2,3] == ([],[1,2,3])

It is equivalent to (take n xs, drop n xs) when n is not _|_ (splitAt _|_ xs = _|_). splitAt is an instance of the more general genericSplitAt, in which n may be of any integral type.

toStreamD :: (Prim a, Monad m) => Array a -> Stream m a

toStreamDRev :: (Prim a, Monad m) => Array a -> Stream m a

toStreamK :: (IsStream t, Prim a) => Array a -> t m a

toStreamKRev :: (IsStream t, Prim a) => Array a -> t m a

toList :: Prim a => Array a -> [a]

Convert an Array into a list.

Internal

writeN :: (MonadIO m, Prim a) => Int -> Fold m a (Array a)

writeN n folds a maximum of n elements from the input stream to an Array.

Internal

data ArrayUnsafe a

Constructors

ArrayUnsafe !(Array a) !Int

writeNUnsafe :: (MonadIO m, Prim a) => Int -> Fold m a (Array a)

Like writeN but does not check the array bounds when writing. The fold driver must not call the step function more than n times otherwise it will corrupt the memory and crash. This function exists mainly because any conditional in the step function blocks fusion causing 10x performance slowdown.

Internal

write :: (MonadIO m, Prim a) => Fold m a (Array a)

Fold the whole input to a single array.

Caution! Do not use this on infinite streams.

Internal

unlines :: (MonadIO m, Prim a) => a -> Stream m (Array a) -> Stream m a