Streamly.Internal.Data.Binary.Decode

Description

Parsers for binary encoded basic Haskell data types.

Documentation

unit :: MonadCatch m => Parser m Word8 ()

A value of type () is encoded as 0 in binary encoding.

0 ==> ()

Internal

bool :: MonadCatch m => Parser m Word8 Bool

A value of type Bool is encoded as follows in binary encoding.

0 ==> False
1 ==> True

Internal

ordering :: MonadCatch m => Parser m Word8 Ordering

A value of type Ordering is encoded as follows in binary encoding.

0 ==> LT
1 ==> EQ
2 ==> GT

Internal

eqWord8 :: MonadCatch m => Word8 -> Parser m Word8 Word8

Accept the input byte only if it is equal to the specified value.

Internal

word8 :: MonadCatch m => Parser m Word8 Word8

Accept any byte.

Internal

word16be :: MonadCatch m => Parser m Word8 Word16

Parse two bytes as a Word16, the first byte is the MSB of the Word16 and second byte is the LSB (big endian representation).

Internal

word16le :: MonadCatch m => Parser m Word8 Word16

Parse two bytes as a Word16, the first byte is the LSB of the Word16 and second byte is the MSB (little endian representation).

Internal

word32be :: MonadCatch m => Parser m Word8 Word32

Parse four bytes as a Word32, the first byte is the MSB of the Word32 and last byte is the LSB (big endian representation).

Internal

word32le :: MonadCatch m => Parser m Word8 Word32

Parse four bytes as a Word32, the first byte is the MSB of the Word32 and last byte is the LSB (big endian representation).

Internal

word64be :: MonadCatch m => Parser m Word8 Word64

Parse eight bytes as a Word64, the first byte is the MSB of the Word64 and last byte is the LSB (big endian representation).

Internal

word64le :: MonadCatch m => Parser m Word8 Word64

Parse eight bytes as a Word64, the first byte is the MSB of the Word64 and last byte is the LSB (big endian representation).

Internal

word64host :: (MonadIO m, MonadCatch m) => Parser m Word8 Word64

Parse eight bytes as a Word64 in the host byte order.

Internal