Streamly.Internal.Data.Either.Strict

Description

| Strict data types to be used as accumulator for strict left folds and scans. For more comprehensive strict data types see https://hackage.haskell.org/package/strict-base-types . The names have been suffixed by a prime so that programmers can easily distinguish the strict versions from the lazy ones.

One major advantage of strict data structures as accumulators in folds and scans is that it helps the compiler optimize the code much better by unboxing. In a big tight loop the difference could be huge.

Documentation

data Either' a b

A strict Either

Constructors

Left' !a

Right' !b