Streamly.Internal.Data.Fold

Description

See Streamly.Data.Fold for an overview.

IMPORTANT: keep the signatures consistent with the folds in Streamly.Prelude

Fold Type

data Step s b

Represents the result of the step of a Fold. Partial returns an intermediate state of the fold, the fold step can be called again with the state or the driver can use extract on the state to get the result out. Done returns the final result and the fold cannot be driven further.

Internal

Constructors

Partial !s

Done !b

data Fold m a b

The type Fold m a b having constructor Fold step initial extract represents a left fold over an input stream of values of type a to a single value of type b in Monad m. The constructor is not exposed via exposed modules, smart constructors are provided to create folds.

The fold uses an intermediate state s as accumulator, the type s is specific to the fold. The initial value of the fold state s is returned by initial. The step function consumes an input and either returns the final result b if the fold is done or the next intermediate state (see Step). At any point the fold driver can extract the result from the intermediate state using the extract function.

Since: 0.7.0

Constructors

forall s. Fold (s -> a -> m (Step s b)) (m (Step s b)) (s -> m b)
Fold step initial extract

Creating

mkAccum :: Monad m => (s -> a -> s) -> s -> (s -> b) -> Fold m a b

Make an accumulating (non-terminating) fold using a pure step function, a pure initial state and a pure state extraction function.

Internal

mkAccum_ :: Monad m => (b -> a -> b) -> b -> Fold m a b

Similar to mkAccum but the final state extracted is identical to the intermediate state.

Internal

mkAccumM :: Functor m => (s -> a -> m s) -> m s -> (s -> m b) -> Fold m a b

Make an accumulating (non-terminating) fold with an effectful step function, an initial state, and a state extraction function.

Internal

mkAccumM_ :: Monad m => (b -> a -> m b) -> m b -> Fold m a b

Similar to mkAccumM but the final state extracted is identical to the intermediate state.

Internal

mkFold :: Monad m => (s -> a -> Step s b) -> Step s b -> (s -> b) -> Fold m a b

Make a terminating fold using a pure step function, a pure initial state and a pure state extraction function.

Internal

mkFold_ :: Monad m => (b -> a -> Step b b) -> Step b b -> Fold m a b

Similar to mkFold but the final state extracted is identical to the intermediate state.

Internal

mkFoldM :: (s -> a -> m (Step s b)) -> m (Step s b) -> (s -> m b) -> Fold m a b

Make a terminating fold with an effectful step function and initial state, and a state extraction function.

mkFoldM = Fold

We can just use Fold but it is provided for completeness.

Internal

mkFoldM_ :: Monad m => (b -> a -> m (Step b b)) -> m (Step b b) -> Fold m a b

Similar to mkFoldM but the final state extracted is identical to the intermediate state.

Internal

Generators

yield :: Applicative m => b -> Fold m a b

A fold that always yields a pure value without consuming any input.

Unimplemented

yieldM :: Applicative m => m b -> Fold m a b

A fold that always yields the result of an effectful action without consuming any input.

Unimplemented

Accumulators

Semigroups and Monoids

sconcat :: (Monad m, Semigroup a) => a -> Fold m a a

Append the elements of an input stream to a provided starting value.

S.fold (FL.sconcat 10) (S.map Sum $ S.enumerateFromTo 1 10)

Internal

mconcat :: (Monad m, Monoid a) => Fold m a a

Fold an input stream consisting of monoidal elements using mappend and mempty.

S.fold FL.mconcat (S.map Sum $ S.enumerateFromTo 1 10)

Since: 0.7.0

foldMap :: (Monad m, Monoid b) => (a -> b) -> Fold m a b

foldMap f = map f mconcat

Make a fold from a pure function that folds the output of the function using mappend and mempty.

S.fold (FL.foldMap Sum) $ S.enumerateFromTo 1 10

Since: 0.7.0

foldMapM :: (Monad m, Monoid b) => (a -> m b) -> Fold m a b

foldMapM f = lmapM f mconcat

Make a fold from a monadic function that folds the output of the function using mappend and mempty.

S.fold (FL.foldMapM (return . Sum)) $ S.enumerateFromTo 1 10

Since: 0.7.0

Reducers

drain :: Monad m => Fold m a ()

A fold that drains all its input, running the effects and discarding the results.

drain = drainBy (const (return ()))

Since: 0.7.0

drainBy :: Monad m => (a -> m b) -> Fold m a ()

drainBy f = lmapM f drain
drainBy = FL.foldMapM (void . f)

Drain all input after passing it through a monadic function. This is the dual of mapM_ on stream producers.

Since: 0.7.0

drainBy2 :: Monad m => (a -> m b) -> Fold2 m c a ()

last :: Monad m => Fold m a (Maybe a)

Extract the last element of the input stream, if any.

last = fmap getLast $ FL.foldMap (Last . Just)

Since: 0.7.0

length :: Monad m => Fold m a Int

Determine the length of the input stream.

length = fmap getSum $ foldMap (Sum . const  1)

Since: 0.7.0

mean :: (Monad m, Fractional a) => Fold m a a

Compute a numerically stable arithmetic mean of all elements in the input stream.

Since: 0.7.0

variance :: (Monad m, Fractional a) => Fold m a a

Compute a numerically stable (population) variance over all elements in the input stream.

Since: 0.7.0

stdDev :: (Monad m, Floating a) => Fold m a a

Compute a numerically stable (population) standard deviation over all elements in the input stream.

Since: 0.7.0

rollingHash :: (Monad m, Enum a) => Fold m a Int64

Compute an Int sized polynomial rolling hash of a stream.

rollingHash = rollingHashWithSalt defaultSalt

Since: 0.7.0

rollingHashWithSalt :: (Monad m, Enum a) => Int64 -> Fold m a Int64

Compute an Int sized polynomial rolling hash

H = salt * k ^ n + c1 * k ^ (n - 1) + c2 * k ^ (n - 2) + ... + cn * k ^ 0

Where c1, c2, cn are the elements in the input stream and k is a constant.

This hash is often used in Rabin-Karp string search algorithm.

See https://en.wikipedia.org/wiki/Rolling_hash

Since: 0.7.0

rollingHashFirstN :: (Monad m, Enum a) => Int -> Fold m a Int64

Compute an Int sized polynomial rolling hash of the first n elements of a stream.

rollingHashFirstN = takeLE n rollingHash

Saturating Reducers

product terminates if it becomes 0. Other folds can theoretically saturate on bounded types, and therefore terminate, however, they will run forever on unbounded types like Integer/Double.

sum :: (Monad m, Num a) => Fold m a a

Determine the sum of all elements of a stream of numbers. Returns additive identity (0) when the stream is empty. Note that this is not numerically stable for floating point numbers.

sum = fmap getSum $ FL.foldMap Sum

Since: 0.7.0

product :: (Monad m, Num a, Eq a) => Fold m a a

Determine the product of all elements of a stream of numbers. Returns multiplicative identity (1) when the stream is empty. The fold terminates when it encounters (0) in its input.

product = fmap getProduct $ FL.foldMap Product

Since 0.8.0 (Added Eq constraint)

Since: 0.7.0

maximumBy :: Monad m => (a -> a -> Ordering) -> Fold m a (Maybe a)

Determine the maximum element in a stream using the supplied comparison function.

Since: 0.7.0

maximum :: (Monad m, Ord a) => Fold m a (Maybe a)

maximum = maximumBy compare

Determine the maximum element in a stream.

Compare with FL.foldMap Max.

Since: 0.7.0

minimumBy :: Monad m => (a -> a -> Ordering) -> Fold m a (Maybe a)

Computes the minimum element with respect to the given comparison function

Since: 0.7.0

minimum :: (Monad m, Ord a) => Fold m a (Maybe a)

Determine the minimum element in a stream using the supplied comparison function.

minimum = minimumBy compare

Compare with FL.foldMap Min.

Since: 0.7.0

Collectors

toList :: Monad m => Fold m a [a]

Folds the input stream to a list.

Warning! working on large lists accumulated as buffers in memory could be very inefficient, consider using Streamly.Data.Array.Foreign instead.

Since: 0.7.0

toListRevF :: Monad m => Fold m a [a]

Buffers the input stream to a list in the reverse order of the input.

Warning! working on large lists accumulated as buffers in memory could be very inefficient, consider using Streamly.Array instead.

Since: 0.7.0

This is more efficient than toList. toList is exactly the same as reversing the list after toListRevF.

Terminating Folds

drainN :: Monad m => Int -> Fold m a ()

A fold that drains the first n elements of its input, running the effects and discarding the results.

index :: Monad m => Int -> Fold m a (Maybe a)

Lookup the element at the given index.

Since: 0.7.0

head :: Monad m => Fold m a (Maybe a)

Extract the first element of the stream, if any.

head = fmap getFirst $ FL.foldMap (First . Just)

Since: 0.7.0

find :: Monad m => (a -> Bool) -> Fold m a (Maybe a)

Returns the first element that satisfies the given predicate.

find p = fmap getFirst $
    FL.foldMap (x -> First (if p x then Just x else Nothing))

Since: 0.7.0

lookup :: (Eq a, Monad m) => a -> Fold m (a, b) (Maybe b)

In a stream of (key-value) pairs (a, b), return the value b of the first pair where the key equals the given value a.

lookup = snd <$> find ((==) . fst)

Since: 0.7.0

findIndex :: Monad m => (a -> Bool) -> Fold m a (Maybe Int)

Returns the first index that satisfies the given predicate.

Since: 0.7.0

elemIndex :: (Eq a, Monad m) => a -> Fold m a (Maybe Int)

Returns the first index where a given value is found in the stream.

elemIndex a = findIndex (== a)

Since: 0.7.0

null :: Monad m => Fold m a Bool

Return True if the input stream is empty.

null = fmap isJust head

Since: 0.7.0

elem :: (Eq a, Monad m) => a -> Fold m a Bool

Return True if the given element is present in the stream.

elem a = any (== a)

Since: 0.7.0

notElem :: (Eq a, Monad m) => a -> Fold m a Bool

Returns True if the given element is not present in the stream.

notElem a = all (/= a)

Since: 0.7.0

all :: Monad m => (a -> Bool) -> Fold m a Bool

Returns True if all elements of a stream satisfy a predicate.

>>> Stream.fold (Fold.all (== 0)) $ Stream.fromList [1,0,1]
> False

Since: 0.7.0

any :: Monad m => (a -> Bool) -> Fold m a Bool

Returns True if any of the elements of a stream satisfies a predicate.

>>> Stream.fold (Fold.any (== 0)) $ Stream.fromList [1,0,1]
> True

Since: 0.7.0

and :: Monad m => Fold m Bool Bool

Returns True if all elements are True, False otherwise

and = all (== True)
and = fmap getAll . FL.foldMap All

Since: 0.7.0

or :: Monad m => Fold m Bool Bool

Returns True if any element is True, False otherwise

or = any (== True)
or = fmap getAny . FL.foldMap Any

Since: 0.7.0

Adapting

hoist :: (forall x. m x -> n x) -> Fold m a b -> Fold n a b

Change the underlying monad of a fold

Internal

generally :: Monad m => Fold Identity a b -> Fold m a b

Adapt a pure fold to any monad

generally = hoist (return . runIdentity)

Internal

Running Incrementally

initialize :: Monad m => Fold m a b -> m (Fold m a b)

Run the initialization effect of a fold. The returned fold would use the value returned by this effect as its initial value.

runStep :: Monad m => Fold m a b -> a -> m (Fold m a b)

Run one step of a fold and store the accumulator as an initial value in the returned fold.

Output Transformations

rsequence :: Monad m => Fold m a (m b) -> Fold m a b

Flatten the monadic output of a fold to pure output.

Since: 0.8.0

sequence :: Monad m => Fold m a (m b) -> Fold m a b

Deprecated: Use rsequence instead

Flatten the monadic output of a fold to pure output.

Since: 0.7.0

rmapM :: Monad m => (b -> m c) -> Fold m a b -> Fold m a c

Map a monadic function on the output of a fold.

Since: 0.8.0

mapM :: Monad m => (b -> m c) -> Fold m a b -> Fold m a c

Deprecated: Use rmapM instead

Map a monadic function on the output of a fold.

Since: 0.7.0

Input Transformations

Mapping

transform :: Monad m => Pipe m a b -> Fold m b c -> Fold m a c

Apply a transformation on a Fold using a Pipe.

Since: 0.7.0

map :: (a -> b) -> Fold m b r -> Fold m a r

(map f fold) maps the function f on the input of the fold.

>>> S.fold (FL.map (\x -> x * x) FL.sum) (S.enumerateFromTo 1 100)
338350

Note: This is not the same as fmap. map is contravariant where as fmap is covariant.

Since: 0.7.0

lmapM :: Monad m => (a -> m b) -> Fold m b r -> Fold m a r

(lmapM f fold) maps the monadic function f on the input of the fold.

Since: 0.7.0

Filtering

filter :: Monad m => (a -> Bool) -> Fold m a r -> Fold m a r

Include only those elements that pass a predicate.

>>> S.fold (filter (> 5) FL.sum) [1..10]
40

Since: 0.7.0

filterM :: Monad m => (a -> m Bool) -> Fold m a r -> Fold m a r

Like filter but with a monadic predicate.

Since: 0.7.0

lcatMaybes :: Monad m => Fold m a b -> Fold m (Maybe a) b

Transform a fold from a pure input to a Maybe input, consuming only Just values.

mapMaybe :: Monad m => (a -> Maybe b) -> Fold m b r -> Fold m a r

mapMaybe f fold maps a Maybe returning function f on the input of the fold, filters out Nothing elements, and return the values extracted from Just.

>>> f x = if even x then Just x else Nothing
>>> fld = Fold.mapMaybe f Fold.toList
>>> Stream.fold fld (Stream.enumerateFromTo 1 10)
[2,4,6,8]

Internal

Trimming

takeLE :: Monad m => Int -> Fold m a b -> Fold m a b

Take at most n input elements and fold them using the supplied fold.

>>> Stream.fold (Fold.takeLE 1 Fold.toList) $ Stream.fromList [1]
[1]

>>> Stream.fold (Fold.takeLE (-1) Fold.toList) $ Stream.fromList [1]
[]

Internal

Since: 0.7.0

takeByTime :: MonadAsync m => Double -> Fold m a b -> Fold m a b

takeByTime n fold uses fold to fold the input items arriving within a window of first n seconds.

Stops when fold stops or when the timeout occurs.

Internal

sliceSepBy :: Monad m => (a -> Bool) -> Fold m a b -> Fold m a b

Consume the input until it encounters an infixed separator element (i.e. when the supplied predicate succeeds), dropping the separator.

Repeated applications of sliceSepBy splits the stream on separator elements determined by the supplied predicate, separator is considered as infixed between two segments, if one side of the separator is missing then it is parsed as an empty stream. The supplied Fold is applied on the split segments. With - representing non-separator elements and . as separator, repeated sliceSepBy splits the stream as follows:

"--.--" => "--" "--"
"--."   => "--" ""
".--"   => ""   "--"

Repeated applications of Fold.sliceSepBy (== x) on the input stream gives us an inverse of Stream.intercalate (Stream.yield x)

Stream.splitOn pred f = Stream.foldMany (Fold.sliceSepBy pred f)

Let's use the following definition for illustration:

splitOn p = Stream.foldMany (Fold.sliceSepBy pred Fold.toList)
splitOn' p = Stream.toList . splitOn p . Stream.fromList

>>> splitOn' (== '.') ""
[""]

>>> splitOn' (== '.') "."
["",""]

>>> splitOn' (== '.') ".a"
> ["","a"]

>>> splitOn' (== '.') "a."
> ["a",""]

>>> splitOn' (== '.') "a.b"
> ["a","b"]

>>> splitOn' (== '.') "a..b"
> ["a","","b"]

Stops - when the predicate succeeds.

Internal

sliceSepByMax :: Monad m => (a -> Bool) -> Int -> Fold m a b -> Fold m a b

Like sliceSepBy but terminates a parse even before the separator is encountered if its size exceeds the specified maximum limit.

take n = PR.sliceSepByMax (const True) n
sliceSepBy p = PR.sliceSepByMax p maxBound

Let's use the following definitions for illustration:

splitOn p n = PR.many FL.toList $ PR.sliceSepByMax p n (FL.toList)
splitOn' p n = S.parse (splitOn p n) . S.fromList

>>> splitOn' (== '.') 0 ""
[""]

>>> splitOn' (== '.') 0 "a"
infinite list of empty strings

>>> splitOn' (== '.') 3 "hello.world"
["hel","lo","wor","ld"]

If the separator is found and the limit is reached at the same time then it behaves just like sliceSepBy i.e. the separator is dropped.

>>> splitOn' (== '.') 0 "."
["",""]

>>> splitOn' (== '.') 0 ".."
["","",""]

Stops - when the predicate succeeds or the limit is reached.

Internal

sliceEndWith :: Monad m => (a -> Bool) -> Fold m a b -> Fold m a b

Collect stream elements until an element succeeds the predicate. Also take the element on which the predicate succeeded. The succeeding element is treated as a suffix separator which is kept in the output segement.

Stops - when the predicate succeeds.

Stream.splitWithSuffix pred f = Stream.foldMany (Fold.sliceEndWith pred f)

Internal

Splitting

splitAt :: Monad m => Int -> Fold m a b -> Fold m a c -> Fold m a (b, c)

splitAt n f1 f2 composes folds f1 and f2 such that first n elements of its input are consumed by fold f1 and the rest of the stream is consumed by fold f2.

let splitAt_ n xs = S.fold (FL.splitAt n FL.toList FL.toList) $ S.fromList xs

>>> splitAt_ 6 "Hello World!"
> ("Hello ","World!")

>>> splitAt_ (-1) [1,2,3]
> ([],[1,2,3])

>>> splitAt_ 0 [1,2,3]
> ([],[1,2,3])

>>> splitAt_ 1 [1,2,3]
> ([1],[2,3])

>>> splitAt_ 3 [1,2,3]
> ([1,2,3],[])

>>> splitAt_ 4 [1,2,3]
> ([1,2,3],[])

splitAt n f1 f2 = splitWith (,) (takeLE n f1) f2

Internal

span :: Monad m => (a -> Bool) -> Fold m a b -> Fold m a c -> Fold m a (b, c)

span p f1 f2 composes folds f1 and f2 such that f1 consumes the input as long as the predicate p is True. f2 consumes the rest of the input.

let span_ p xs = S.fold (FL.span p FL.toList FL.toList) $ S.fromList xs

>>> span_ (< 1) [1,2,3]
> ([],[1,2,3])

>>> span_ (< 2) [1,2,3]
> ([1],[2,3])

>>> span_ (< 4) [1,2,3]
> ([1,2,3],[])

Internal

break :: Monad m => (a -> Bool) -> Fold m a b -> Fold m a c -> Fold m a (b, c)

break p = span (not . p)

Break as soon as the predicate becomes True. break p f1 f2 composes folds f1 and f2 such that f1 stops consuming input as soon as the predicate p becomes True. The rest of the input is consumed f2.

This is the binary version of splitBy.

let break_ p xs = S.fold (S.break p FL.toList FL.toList) $ S.fromList xs

>>> break_ (< 1) [3,2,1]
> ([3,2,1],[])

>>> break_ (< 2) [3,2,1]
> ([3,2],[1])

>>> break_ (< 4) [3,2,1]
> ([],[3,2,1])

Internal

spanBy :: Monad m => (a -> a -> Bool) -> Fold m a b -> Fold m a c -> Fold m a (b, c)

Break the input stream into two groups, the first group takes the input as long as the predicate applied to the first element of the stream and next input element holds True, the second group takes the rest of the input.

Internal

spanByRolling :: Monad m => (a -> a -> Bool) -> Fold m a b -> Fold m a c -> Fold m a (b, c)

Like spanBy but applies the predicate in a rolling fashion i.e. predicate is applied to the previous and the next input elements.

Internal

Distributing

tee :: Monad m => Fold m a b -> Fold m a c -> Fold m a (b, c)

Distribute one copy of the stream to each fold and zip the results.

                |-------Fold m a b--------|
---stream m a---|                         |---m (b,c)
                |-------Fold m a c--------|

>>> S.fold (FL.tee FL.sum FL.length) (S.enumerateFromTo 1.0 100.0)
(5050.0,100)

Since: 0.7.0

teeWith :: Monad m => (a -> b -> c) -> Fold m x a -> Fold m x b -> Fold m x c

teeWith k f1 f2 distributes its input to both f1 and f2 until both of them terminate and combines their output using k.

Internal

teeWithFst :: (b -> c -> d) -> Fold m a b -> Fold m a c -> Fold m a d

Like teeWith but terminates when the first fold terminates.

Unimplemented

teeWithMin :: (b -> c -> d) -> Fold m a b -> Fold m a c -> Fold m a d

Like teeWith but terminates when any fold terminates.

Unimplemented

distribute :: Monad m => [Fold m a b] -> Fold m a [b]

Distribute one copy of the stream to each fold and collect the results in a container.

                |-------Fold m a b--------|
---stream m a---|                         |---m [b]
                |-------Fold m a b--------|
                |                         |
                           ...

>>> S.fold (FL.distribute [FL.sum, FL.length]) (S.enumerateFromTo 1 5)
[15,5]

This is the consumer side dual of the producer side sequence operation.

Stops when all the folds stop.

Since: 0.7.0

Partitioning

partitionByM :: Monad m => (a -> m (Either b c)) -> Fold m b x -> Fold m c y -> Fold m a (x, y)

Partition the input over two folds using an Either partitioning predicate.

                                    |-------Fold b x--------|
-----stream m a --> (Either b c)----|                       |----(x,y)
                                    |-------Fold c y--------|

Send input to either fold randomly:

>>> import System.Random (randomIO)
>>> randomly a = randomIO >>= \x -> return $ if x then Left a else Right a
>>> S.fold (FL.partitionByM randomly FL.length FL.length) (S.enumerateFromTo 1 100)
(59,41)

Send input to the two folds in a proportion of 2:1:

import Data.IORef (newIORef, readIORef, writeIORef)
proportionately m n = do
 ref <- newIORef $ cycle $ concat [replicate m Left, replicate n Right]
 return $ \a -> do
     r <- readIORef ref
     writeIORef ref $ tail r
     return $ head r a

main = do
 f <- proportionately 2 1
 r <- S.fold (FL.partitionByM f FL.length FL.length) (S.enumerateFromTo (1 :: Int) 100)
 print r

(67,33)

This is the consumer side dual of the producer side mergeBy operation.

When one fold is done, any input meant for it is ignored until the other fold is also done.

Stops when both the folds stop.

See also: partitionByFstM and partitionByMinM.

Since: 0.7.0

partitionByFstM :: (a -> m (Either b c)) -> Fold m b x -> Fold m c y -> Fold m a (x, y)

Similar to partitionByM but terminates when the first fold terminates.

Unimplemented

partitionByMinM :: (a -> m (Either b c)) -> Fold m b x -> Fold m c y -> Fold m a (x, y)

Similar to partitionByM but terminates when any fold terminates.

Unimplemented

partitionBy :: Monad m => (a -> Either b c) -> Fold m b x -> Fold m c y -> Fold m a (x, y)

Same as partitionByM but with a pure partition function.

Count even and odd numbers in a stream:

>>> let f = FL.partitionBy (\n -> if even n then Left n else Right n)
                      (fmap (("Even " ++) . show) FL.length)
                      (fmap (("Odd "  ++) . show) FL.length)
  in S.fold f (S.enumerateFromTo 1 100)
("Even 50","Odd 50")

Since: 0.7.0

partition :: Monad m => Fold m b x -> Fold m c y -> Fold m (Either b c) (x, y)

Compose two folds such that the combined fold accepts a stream of Either and routes the Left values to the first fold and Right values to the second fold.

partition = partitionBy id

Since: 0.7.0

Demultiplexing

demux :: (Monad m, Ord k) => Map k (Fold m a b) -> Fold m (k, a) (Map k b)

Fold a stream of key value pairs using a map of specific folds for each key into a map from keys to the results of fold outputs of the corresponding values.

> let table = Data.Map.fromList [("SUM", FL.sum), ("PRODUCT", FL.product)]
      input = S.fromList [("SUM",1),("PRODUCT",2),("SUM",3),("PRODUCT",4)]
  in S.fold (FL.demux table) input
fromList [(PRODUCT,8),(SUM,4)]

Since: 0.7.0

demuxWith :: (Monad m, Ord k) => (a -> (k, a')) -> Map k (Fold m a' b) -> Fold m a (Map k b)

Split the input stream based on a key field and fold each split using a specific fold collecting the results in a map from the keys to the results. Useful for cases like protocol handlers to handle different type of packets using different handlers.

                            |-------Fold m a b
-----stream m a-----Map-----|
                            |-------Fold m a b
                            |
                                      ...

Any input that does not map to a fold in the input Map is silently ignored.

demuxWith f kv = fmap fst $ demuxDefaultWith f kv drain

Internal

demuxDefault :: (Monad m, Ord k) => Map k (Fold m a b) -> Fold m (k, a) b -> Fold m (k, a) (Map k b, b)

demuxDefaultWith :: (Monad m, Ord k) => (a -> (k, a')) -> Map k (Fold m a' b) -> Fold m (k, a') c -> Fold m a (Map k b, c)

Like demuxWith but uses a default catchall fold to handle inputs which do not have a specific fold in the map to handle them.

If any fold in the map stops, inputs meant for that fold are sent to the catchall fold. If the catchall fold stops then inputs that do not match any fold are ignored.

Stops when all the folds, including the catchall fold, stop.

Internal

Classifying

classify :: (Monad m, Ord k) => Fold m a b -> Fold m (k, a) (Map k b)

Given an input stream of key value pairs and a fold for values, fold all the values belonging to each key. Useful for map/reduce, bucketizing the input in different bins or for generating histograms.

> let input = S.fromList [("ONE",1),("ONE",1.1),("TWO",2), ("TWO",2.2)]
  in S.fold (FL.classify FL.toList) input
fromList [("ONE",[1.1,1.0]),("TWO",[2.2,2.0])]

Since: 0.7.0

classifyWith :: (Monad m, Ord k) => (a -> k) -> Fold m a b -> Fold m a (Map k b)

Split the input stream based on a key field and fold each split using the given fold. Useful for map/reduce, bucketizing the input in different bins or for generating histograms.

> let input = S.fromList [("ONE",1),("ONE",1.1),("TWO",2), ("TWO",2.2)]
  in S.fold (FL.classify FL.toList) input
fromList [("ONE",[1.1,1.0]),("TWO",[2.2,2.0])]

If the classifier fold stops for a particular key any further inputs in that bucket are ignored.

Stops: never

Internal

Unzipping

unzip :: Monad m => Fold m a x -> Fold m b y -> Fold m (a, b) (x, y)

Send the elements of tuples in a stream of tuples through two different folds.

                          |-------Fold m a x--------|
---------stream of (a,b)--|                         |----m (x,y)
                          |-------Fold m b y--------|

This is the consumer side dual of the producer side zip operation.

Since: 0.7.0

unzipWith :: Monad m => (a -> (b, c)) -> Fold m b x -> Fold m c y -> Fold m a (x, y)

Split elements in the input stream into two parts using a pure splitter function, direct each part to a different fold and zip the results.

unzipWith f fld1 fld2 = map f (unzip fld1 fld2)

This fold terminates when both the input folds terminate.

Since: 0.7.0

unzipWithM :: Monad m => (a -> m (b, c)) -> Fold m b x -> Fold m c y -> Fold m a (x, y)

Like unzipWith but with a monadic splitter function.

- unzipWithM k f1 f2 = lmapM k (unzip f1 f2)

Since: 0.7.0

unzipWithFstM :: (a -> m (b, c)) -> Fold m b x -> Fold m c y -> Fold m a (x, y)

Similar to unzipWithM but terminates when the first fold terminates.

Unimplemented

unzipWithMinM :: (a -> m (b, c)) -> Fold m b x -> Fold m c y -> Fold m a (x, y)

Similar to unzipWithM but terminates when any fold terminates.

Unimplemented

Nesting

many :: Monad m => Fold m b c -> Fold m a b -> Fold m a c

Collect zero or more applications of a fold. many collect split applies the split fold repeatedly on the input stream and accumulates zero or more fold results using collect.

Stops when collect stops.

Internal

See also: Streamly.Prelude.concatMap, Streamly.Prelude.foldMany

intervalsOf :: MonadAsync m => Double -> Fold m a b -> Fold m b c -> Fold m a c

Group the input stream into windows of n second each and then fold each group using the provided fold function.

For example, we can copy and distribute a stream to multiple folds where each fold can group the input differently e.g. by one second, one minute and one hour windows respectively and fold each resulting stream of folds.

-----Fold m a b----|-Fold n a c-|-Fold n a c-|-...-|----Fold m a c

intervalsOf n split collect = many collect (takeByTime n split)

Internal

chunksOf :: Monad m => Int -> Fold m a b -> Fold m b c -> Fold m a c

chunksOf n split collect repeatedly applies the split fold to chunks of n items in the input stream and supplies the result to the collect fold.

chunksOf n split collect = many collect (takeLE n split)

Stops when collect stops.

Internal

chunksBetween :: Int -> Int -> Fold m a b -> Fold m b c -> Fold m a c

Group the input stream into groups of elements between low and high. Collection starts in chunks of low and then keeps doubling until we reach high. Each chunk is folded using the provided fold function.

This could be useful, for example, when we are folding a stream of unknown size to a stream of arrays and we want to minimize the number of allocations.

NOTE: this would be an application of "many" using a terminating fold.

Unimplemented

concatSequence :: Fold m b c -> t (Fold m a b) -> Fold m a c

concatSequence f t applies folds from stream t sequentially and collects the results using the fold f.

Unimplemented

concatMap :: (b -> Fold m a c) -> Fold m a b -> Fold m a c

Map a Fold returning function on the result of a Fold.

Compare with Monad instance method >>=. This implementation allows stream fusion but has quadratic complexity. This can fuse with other operations and can be much faster than Monad instance for small number (less than 8) of compositions.

Unimplemented

duplicate :: Monad m => Fold m a b -> Fold m a (Fold m a b)

Modify the fold such that it returns a new Fold instead of the output. If the fold was already done the returned fold would always yield the result. If the fold was partial, the returned fold starts from where we left i.e. it uses the last accumulator value as the initial value of the accumulator. Thus we can resume the fold later and feed it more input.

> do
    more <- S.fold (FL.duplicate FL.sum) (S.enumerateFromTo 1 10)
    evenMore <- S.fold (FL.duplicate more) (S.enumerateFromTo 11 20)
    S.fold evenMore (S.enumerateFromTo 21 30)
 465

Since: 0.7.0