Streamly.Internal.Data.Fold.Step

Description

Types

data Step s b

Represents the result of the step of a Fold. Partial returns an intermediate state of the fold, the fold step can be called again with the state or the driver can use extract on the state to get the result out. Done returns the final result and the fold cannot be driven further.

Pre-release

Constructors

Partial !s

Done !b

mapMStep :: Applicative m => (a -> m b) -> Step s a -> m (Step s b)

Map a monadic function over the result b in Step s b.

Internal

chainStepM :: Applicative m => (s1 -> m s2) -> (a -> m (Step s2 b)) -> Step s1 a -> m (Step s2 b)

If Partial then map the state, if Done then call the next step.