Streamly.Internal.Data.Fold.Tee

Description

Documentation

newtype Tee m a b

The type Tee m a b represents a left fold over an input stream of values of type a to a single value of type b in Monad m.

Tee is a wrapper over Fold that uses teeWith to define the applicative instance.

Internal

Constructors

Tee

Fields

runTee :: Fold m a b

fromFold :: Fold m a b -> Tee m a b

Convert a Fold to Tee.

toFold :: Tee m a b -> Fold m a b

Convert a Tee to Fold.