Streamly.Internal.Data.IORef.Prim

Description

A mutable variable in a mutation capable monad (IO) holding a Prim value. This allows fast modification because of unboxed storage.

Multithread Consistency Notes

In general, any value that straddles a machine word cannot be guaranteed to be consistently read from another thread without a lock. GHC heap objects are always machine word aligned, therefore, a IORef is also word aligned. On a 64-bit platform, writing a 64-bit aligned type from one thread and reading it from another thread should give consistent old or new value. The same holds true for 32-bit values on a 32-bit platform.

Documentation

data IORef a

An IORef holds a single Prim value.

class Prim a

Class of types supporting primitive array operations. This includes interfacing with GC-managed memory (functions suffixed with ByteArray#) and interfacing with unmanaged memory (functions suffixed with Addr#). Endianness is platform-dependent.

Minimal complete definition

sizeOf#, alignment#, indexByteArray#, readByteArray#, writeByteArray#, setByteArray#, indexOffAddr#, readOffAddr#, writeOffAddr#, setOffAddr#

Construction

newIORef :: forall a. Prim a => a -> IO (IORef a)

Create a new IORef.

Internal

Write

writeIORef :: Prim a => IORef a -> a -> IO ()

Write a value to an IORef.

Internal

modifyIORef' :: Prim a => IORef a -> (a -> a) -> IO ()

Modify the value of an IORef using a function with strict application.

Internal

Read

readIORef :: Prim a => IORef a -> IO a

Read a value from an IORef.

Internal