Streamly.Internal.Data.Parser.ParserD

Description

Direct style parser implementation with stream fusion.

Documentation

data Parser m a b

A parser is a fold that can fail and is represented as Parser step initial extract. Before we drive a parser we call the initial action to retrieve the initial state of the fold. The parser driver invokes step with the state returned by the previous step and the next input element. It results into a new state and a command to the driver represented by Step type. The driver keeps invoking the step function until it stops or fails. At any point of time the driver can call extract to inspect the result of the fold. It may result in an error or an output value.

Internal

Constructors

Parser (s -> a -> m (Step s b)) (m s) (s -> m b)

newtype ParseError

This exception is used for two purposes:

When a parser ultimately fails, the user of the parser is intimated via this exception.

When the "extract" function of a parser needs to throw an error.

Internal

Constructors

ParseError String

data Step s b

The return type of a Parser step.

A parser is driven by a parse driver one step at a time, at any time the driver may extract the result of the parser. The parser may ask the driver to backtrack at any point, therefore, the driver holds the input up to a point of no return in a backtracking buffer. The buffer grows or shrinks based on the return values of the parser step execution.

When a parser step is executed it generates a new intermediate state of the parse result along with a command to the driver. The command tells the driver whether to keep the input stream for a potential backtracking later on or drop it, and how much to keep. The constructors of Step represent the commands to the driver.

Internal

Constructors

Yield Int s
Yield offset state indicates that the parser has yielded a new result which is a point of no return. The result can be extracted using extract. The driver drops the buffer except offset elements before the current position in stream. The rule is that if a parser has yielded at least once it cannot return a failure result.

Skip Int s
Skip offset state indicates that the parser has consumed the current input but no new result has been generated. A new state is generated. However, if we use extract on state it will generate a result from the previous Yield. When offset is non-zero it is a backward offset from the current position in the stream from which the driver will feed the next input to the parser. The offset cannot be beyond the latest point of no return created by Yield.

Stop Int b
Stop offset result asks the driver to stop driving the parser because it has reached a fixed point and further input will not change the result. offset is the count of unused elements which includes the element on which Stop occurred.

Error String
An error makes the parser backtrack to the last checkpoint and try another alternative.

Accumulators

fromFold :: Monad m => Fold m a b -> Parser m a b

See fromFold.

Internal

toParserK :: MonadCatch m => Parser m a b -> Parser m a b

Convert a direct style Parser to a CPS style Parser.

Internal

fromParserK :: Monad m => Parser m a b -> Parser m a b

Convert a CPS style Parser to a direct style Parser.

Unimplemented

any :: Monad m => (a -> Bool) -> Parser m a Bool

all :: Monad m => (a -> Bool) -> Parser m a Bool

yield :: Monad m => b -> Parser m a b

See yield.

Internal

yieldM :: Monad m => m b -> Parser m a b

See yieldM.

Internal

die :: MonadThrow m => String -> Parser m a b

See die.

Internal

dieM :: MonadThrow m => m String -> Parser m a b

See dieM.

Internal

Element parsers

peek :: MonadThrow m => Parser m a a

See peek.

Internal

eof :: Monad m => Parser m a ()

See eof.

Internal

satisfy :: MonadThrow m => (a -> Bool) -> Parser m a a

See satisfy.

Internal

Sequence parsers

take :: Monad m => Int -> Fold m a b -> Parser m a b

See take.

Internal

takeEQ :: MonadThrow m => Int -> Fold m a b -> Parser m a b

See takeEQ.

Internal

takeGE :: MonadThrow m => Int -> Fold m a b -> Parser m a b

See takeGE.

Internal

lookAhead :: MonadThrow m => Parser m a b -> Parser m a b

See lookahead.

Internal

takeWhile :: Monad m => (a -> Bool) -> Fold m a b -> Parser m a b

See takeWhile.

Internal

takeWhile1 :: MonadThrow m => (a -> Bool) -> Fold m a b -> Parser m a b

See takeWhile1.

Internal

sliceSepBy :: Monad m => (a -> Bool) -> Fold m a b -> Parser m a b

See sliceSepBy.

Internal

sliceSepByMax :: Monad m => (a -> Bool) -> Int -> Fold m a b -> Parser m a b

See sliceSepByMax.

Internal

sliceEndWith :: (a -> Bool) -> Fold m a b -> Parser m a b

See sliceEndWith.

Unimplemented

sliceBeginWith :: (a -> Bool) -> Fold m a b -> Parser m a b

See sliceBeginWith.

Unimplemented

wordBy :: (a -> Bool) -> Fold m a b -> Parser m a b

See wordBy.

Unimplemented

groupBy :: (a -> a -> Bool) -> Fold m a b -> Parser m a b

See groupBy.

Unimplemented

eqBy :: MonadThrow m => (a -> a -> Bool) -> [a] -> Parser m a ()

See eqBy.

Internal

Binary Combinators

Sequential Applicative

splitWith :: Monad m => (a -> b -> c) -> Parser m x a -> Parser m x b -> Parser m x c

See splitWith.

Internal

split_ :: Monad m => Parser m x a -> Parser m x b -> Parser m x b

See split_.

Internal

Parallel Applicatives

teeWith :: Monad m => (a -> b -> c) -> Parser m x a -> Parser m x b -> Parser m x c

See teeWith.

Internal

teeWithFst :: Monad m => (a -> b -> c) -> Parser m x a -> Parser m x b -> Parser m x c

See teeWithFst.

Internal

teeWithMin :: (a -> b -> c) -> Parser m x a -> Parser m x b -> Parser m x c

See teeWithMin.

Unimplemented

Sequential Interleaving

deintercalate :: Fold m a y -> Parser m x a -> Fold m b z -> Parser m x b -> Parser m x (y, z)

See deintercalate.

Unimplemented

Sequential Alternative

alt :: Monad m => Parser m x a -> Parser m x a -> Parser m x a

See alt.

Internal

Parallel Alternatives

shortest :: Monad m => Parser m x a -> Parser m x a -> Parser m x a

See shortest.

Internal

longest :: MonadCatch m => Parser m x a -> Parser m x a -> Parser m x a

See longest.

Internal

N-ary Combinators

Sequential Collection

sequence :: Fold m b c -> t (Parser m a b) -> Parser m a c

See sequence.

Unimplemented

concatMap :: Monad m => (b -> Parser m a c) -> Parser m a b -> Parser m a c

See concatMap.

Internal

Sequential Repetition

count :: Int -> Fold m b c -> Parser m a b -> Parser m a c

See count.

Unimplemented

countBetween :: Int -> Int -> Fold m b c -> Parser m a b -> Parser m a c

See countBetween.

Unimplemented

many :: MonadCatch m => Fold m b c -> Parser m a b -> Parser m a c

See many.

Internal

some :: MonadCatch m => Fold m b c -> Parser m a b -> Parser m a c

See some.

Internal

manyTill :: MonadCatch m => Fold m b c -> Parser m a b -> Parser m a x -> Parser m a c

See manyTill.

Internal

Collection of Alternatives

choice :: t (Parser m a b) -> Parser m a b

See choice.

Unimplemented