Streamly.Internal.Data.Parser.ParserK.Type

Description

CPS style implementation of parsers.

The CPS representation allows linear performance for Applicative, sequenceA, Monad, sequence, and Alternative, choice operations compared to the quadratic complexity of the corresponding direct style operations. However, direct style operations allow fusion with ~10x better performance than CPS.

The direct style representation does not allow for recursive definitions of "some" and "many" whereas CPS allows that.

Documentation

newtype Parser m a b

A continuation passing style parser representation.

Constructors

MkParser

Fields

runParser :: forall r. Int -> (Int, Int) -> ((Int, Int) -> Parse b -> m (Driver m a r)) -> m (Driver m a r)

fromPure :: b -> Parser m a b

A parser that always yields a pure value without consuming any input.

Pre-release

fromEffect :: Monad m => m b -> Parser m a b

See fromEffect.

Pre-release

die :: String -> Parser m a b

A parser that always fails with an error message without consuming any input.

Pre-release

Conversion

toParserK :: MonadCatch m => Parser m a b -> Parser m a b

Convert a direct style Parser to a CPS style Parser.

Pre-release

fromParserK :: MonadThrow m => Parser m a b -> Parser m a b

Convert a CPS style Parser to a direct style Parser.

"initial" returns a continuation which can be called one input at a time using the "step" function.

Pre-release