Streamly.Internal.Data.Parser.Types

Description

Streaming and backtracking parsers.

The Parser type or a parsing fold is a generalization of the Fold type. The Fold type always succeeds on each input. Therefore it does not need to buffer the input. In contrast, a Parser may fail and backtrack to replay the input again to explore another branch of the parser. Therefore, it needs to buffer the input.

The Parser type is an extension of the Fold type. The Step functor for the Parser type has been specifically designed for stream parsing applications, assuming the input to be a seuqentially read stream and potentially buffering a contiguous segment of the stream based on the parser's need.

To provide richer communication with the input stream, the Step functor of a Parser provides more powerful commands to manipulate the input stream after each step invocation. It allows the fold driver to:

Buffer the input until explicitly asked to drop the input

Drop the input stream beyond a "point of no return" in the history because the input till that point has been consumed irrevocably.

Go back in the buffered stream upto the point of no return.

Based on application requirements it should be possible to design even a richer interface to manipulate the input stream/buffer. For example, we could randomly seek into the stream in the forward or reverse directions or we can even seek to the end or from the end or seek from the beginning.

Documentation

data Step s

The return type of a Parser step.

A parser is driven by a parse driver one step at a time, at any time the driver may extract the result of the parser. The parser may ask the driver to backtrack at any point, therefore, the driver holds the input up to a point of no return in a backtracking buffer. The buffer grows or shrinks based on the return values of the parser step execution.

When a parser step is executed it generates a new state of the parse result along with a command to the driver. The command tells the driver whether to keep the input stream for backtracking or drop it, and how much to keep. The constructors of Step represent the commands to the driver.

Internal

Constructors

Yield Int s
Yield offset state indicates that the parser has yielded a new result which is a point of no return. The result can be extracted using extract. The driver drops the buffer except offset elements before the current position in stream. The rule is that if a parser has yielded at least once it cannot return a failure result.

Skip Int s
Skip offset state indicates that the parser has consumed the current input but no new result has been generated. A new state is generated. However, if we use extract on state it will generate a result from the previous Yield. When offset is non-zero it is a backward offset from the current position in the stream from which the driver will feed the next input to the parser. The offset cannot be beyond the latest point of no return created by Yield.

Stop Int s
Stop offset state asks the driver to stop driving the parser because it has reached a fixed point and further input will not change the result. offset is the count of unused elements which includes the element on which Stop occurred. Once a fold stops, driving it further may produce undefined behavior.

data Parser m a b

A parsing fold is represented as Parser step initial extract. Before we drive a parser we call the initial action to retrieve the initial state of the fold. The driver invokes step with the state returned by the previous step and the next input element. It results into a new state and a command to the driver represented by Step type. At any point of time the driver can call extract to inspect the result of the fold. It may result in an error or an output value.

Constructors

Parser (s -> a -> m (Step s)) (m s) (s -> m (Either String b))

splitWith :: Monad m => (a -> b -> c) -> Parser m x a -> Parser m x b -> Parser m x c

Apply two parsers sequentially to an input stream. The input is provided to the first parser, when it is done the remaining input is provided to the second parser. If both the parsers succeed there outputs are combined using the supplied function. The operation fails if any of the parsers fail.

This undoes an "append" of two streams, it splits the streams.