Streamly.Internal.Data.ParserK

Description

Parsers.

Documentation

newtype Parser m a b

Constructors

MkParser

Fields

runParser :: forall r. Zipper m a -> ((Zipper m a, Either String b) -> m r) -> m r

any :: MonadCatch m => (a -> Bool) -> Parser m a Bool

>>> S.parse (PR.any (== 0)) $ S.fromList [1,0,1]
> Right True

satisfy :: MonadCatch m => (a -> Bool) -> Parser m a a

Take the element if the predicate succeeds, fail otherwise.

>>> S.parse (PR.satisfy (== 1)) $ S.fromList [1,0,1]
> 1

Internal

choice :: (Foldable t, Alternative f) => t (f a) -> f a

choice ps tries to apply the actions in the list ps in order, until one of them succeeds. Returns the value of the succeeding action.