Streamly.Internal.Data.Prim.Array

Description

Documentation

data PrimArray a

Arrays of unboxed elements. This accepts types like Double, Char, Int, and Word, as well as their fixed-length variants (Word8, Word16, etc.). Since the elements are unboxed, a PrimArray is strict in its elements. This differs from the behavior of Array, which is lazy in its elements.

Constructors

PrimArray ByteArray#

class Prim a where

Methods

sizeOf# :: a -> Int#

alignment# :: a -> Int#

indexByteArray# :: ByteArray# -> Int# -> a

readByteArray# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, a#)

writeByteArray# :: MutableByteArray# s -> Int# -> a -> State# s -> State# s

setByteArray# :: MutableByteArray# s -> Int# -> Int# -> a -> State# s -> State# s

indexOffAddr# :: Addr# -> Int# -> a

readOffAddr# :: Addr# -> Int# -> State# s -> (#State# s, a#)

writeOffAddr# :: Addr# -> Int# -> a -> State# s -> State# s

setOffAddr# :: Addr# -> Int# -> Int# -> a -> State# s -> State# s

foldl' :: Prim a => (b -> a -> b) -> b -> PrimArray a -> b

foldr :: Prim a => (a -> b -> b) -> b -> PrimArray a -> b

length :: Prim a => PrimArray a -> Int

writeN :: (MonadIO m, Prim a) => Int -> Fold m a (PrimArray a)

write :: (MonadIO m, Prim a) => Fold m a (PrimArray a)

toStreamD :: (Prim a, Monad m) => PrimArray a -> Stream m a

toStreamDRev :: (Prim a, Monad m) => PrimArray a -> Stream m a

toStream :: (Prim a, Monad m, IsStream t) => PrimArray a -> t m a

toStreamRev :: (Prim a, Monad m, IsStream t) => PrimArray a -> t m a

read :: (Prim a, Monad m) => Unfold m (PrimArray a) a

readSlice :: (Prim a, Monad m) => Int -> Int -> Unfold m (PrimArray a) a

fromListN :: Prim a => Int -> [a] -> PrimArray a

fromList :: Prim a => [a] -> PrimArray a

fromStreamDN :: (MonadIO m, Prim a) => Int -> Stream m a -> m (PrimArray a)

fromStreamD :: (MonadIO m, Prim a) => Stream m a -> m (PrimArray a)

fromStreamN :: (MonadIO m, Prim a) => Int -> SerialT m a -> m (PrimArray a)

fromStream :: (MonadIO m, Prim a) => SerialT m a -> m (PrimArray a)

streamFold :: (Prim a, Monad m) => (SerialT m a -> m b) -> PrimArray a -> m b

fold :: (Prim a, Monad m) => Fold m a b -> PrimArray a -> m b

Orphan instances

Prim a => NFData (PrimArray a)

Instance details

Methods

rnf :: PrimArray a -> ()