Streamly.Internal.Data.Producer.Type

Description

See Streamly.Internal.Data.Producer for introduction.

Type

data Producer m a b

A Producer m a b is a generator of a stream of values of type b from a seed of type a in Monad m.

Pre-release

Constructors

forall s. Producer (s -> m (Step s b)) (a -> m s) (s -> m a)
Producer step inject extract

Producers

nil :: Monad m => Producer m a b

nilM :: Monad m => (a -> m c) -> Producer m a b

unfoldrM :: Monad m => (a -> m (Maybe (b, a))) -> Producer m a b

fromList :: Monad m => Producer m [a] a

Convert a list of pure values to a Stream

Pre-release

Mapping

translate :: Functor m => (a -> c) -> (c -> a) -> Producer m c b -> Producer m a b

Interconvert the producer between two interconvertible input types.

Pre-release

lmap :: (a -> a) -> Producer m a b -> Producer m a b

Map the producer input to another value of the same type.

Pre-release

Nesting

data NestedLoop s1 s2

State representing a nested loop.

Constructors

OuterLoop s1

InnerLoop s1 s2

concat :: Monad m => Producer m a b -> Producer m b c -> Producer m (NestedLoop a b) c

Apply the second unfold to each output element of the first unfold and flatten the output in a single stream.

Pre-release