Streamly.Internal.Data.Stream.IsStream.Common

Description

Generation

yield :: IsStream t => a -> t m a

yield a = a `cons` nil

Create a singleton stream from a pure value.

The following holds in monadic streams, but not in Zip streams:

yield = pure
yield = yieldM . pure

In Zip applicative streams yield is not the same as pure because in that case pure is equivalent to repeat instead. yield and pure are equally efficient, in other cases yield may be slightly more efficient than the other equivalent definitions.

Since: 0.4.0

yieldM :: (Monad m, IsStream t) => m a -> t m a

yieldM m = m `consM` nil

Create a singleton stream from a monadic action.

> toList $ yieldM getLine
hello
["hello"]

Since: 0.4.0

repeatM :: (IsStream t, MonadAsync m) => m a -> t m a

repeatM = fix . consM
repeatM = cycle1 . yieldM

Generate a stream by repeatedly executing a monadic action forever.

drain $ serially $ S.take 10 $ S.repeatM $ (threadDelay 1000000 >> print 1)
drain $ asyncly  $ S.take 10 $ S.repeatM $ (threadDelay 1000000 >> print 1)

Concurrent, infinite (do not use with parallely)

Since: 0.2.0

timesWith :: (IsStream t, MonadAsync m) => Double -> t m (AbsTime, RelTime64)

timesWith g returns a stream of time value tuples. The first component of the tuple is an absolute time reference (epoch) denoting the start of the stream and the second component is a time relative to the reference.

The argument g specifies the granularity of the relative time in seconds. A lower granularity clock gives higher precision but is more expensive in terms of CPU usage. Any granularity lower than 1 ms is treated as 1 ms.

>>> S.mapM_ (x -> print x >> threadDelay 1000000) $ S.timesWith 0.01
> (AbsTime (TimeSpec {sec = 2496295, nsec = 536223000}),RelTime64 (NanoSecond64 0))
> (AbsTime (TimeSpec {sec = 2496295, nsec = 536223000}),RelTime64 (NanoSecond64 1002028000))
> (AbsTime (TimeSpec {sec = 2496295, nsec = 536223000}),RelTime64 (NanoSecond64 1996656000))

Note: This API is not safe on 32-bit machines.

Internal

absTimesWith :: (IsStream t, MonadAsync m, Functor (t m)) => Double -> t m AbsTime

absTimesWith g returns a stream of absolute timestamps using a clock of granularity g specified in seconds. A low granularity clock is more expensive in terms of CPU usage. Any granularity lower than 1 ms is treated as 1 ms.

>>> S.mapM_ print $ S.delayPre 1 $ S.absTimesWith 0.01

Note: This API is not safe on 32-bit machines.

Internal

relTimesWith :: (IsStream t, MonadAsync m, Functor (t m)) => Double -> t m RelTime64

relTimesWith g returns a stream of relative time values starting from 0, using a clock of granularity g specified in seconds. A low granularity clock is more expensive in terms of CPU usage. Any granularity lower than 1 ms is treated as 1 ms.

>>> S.mapM_ print $ S.delayPre 1 $ S.relTimesWith 0.01
> RelTime64 (NanoSecond64 0)
> RelTime64 (NanoSecond64 91139000)
> RelTime64 (NanoSecond64 204052000)

Note: This API is not safe on 32-bit machines.

Internal

Elimination

fold :: Monad m => Fold m a b -> SerialT m a -> m b

Fold a stream using the supplied left Fold and reducing the resulting expression strictly at each step. The behavior is similar to foldl'. A Fold can terminate early without consuming the full stream. See the documentation of individual Folds for termination behavior.

>>> S.fold FL.sum (S.enumerateFromTo 1 100)
5050

Since: 0.7.0

Transformation

scanlMAfter' :: (IsStream t, Monad m) => (b -> a -> m b) -> m b -> (b -> m b) -> t m a -> t m b

scanlMAfter' accumulate initial done stream is like scanlM' except that it provides an additional done function to be applied on the accumulator when the stream stops. The result of done is also emitted in the stream.

This function can be used to allocate a resource in the beginning of the scan and release it when the stream ends or to flush the internal state of the scan at the end.

Internal

postscanlM' :: (IsStream t, Monad m) => (b -> a -> m b) -> m b -> t m a -> t m b

Like postscanl' but with a monadic step function and a monadic seed.

Since: 0.7.0

Since: 0.8.0 (signature change)

smapM :: (IsStream t, Monad m) => (s -> a -> m (s, b)) -> m s -> t m a -> t m b

A stateful mapM, equivalent to a left scan, more like mapAccumL. Hopefully, this is a better alternative to scan. Separation of state from the output makes it easier to think in terms of a shared state, and also makes it easier to keep the state fully strict and the output lazy.

See also: scanlM'

Internal

The stateful step function can be simplified to (s -> a -> m b) to provide a read-only environment. However, that would just be mapM.

The initial action could be m (s, Maybe b), and we can also add a final action s -> m (Maybe b). This can be used to get pre/post scan like functionality and also to flush the state in the end like scanlMAfter'. We can also use it along with a fusible version of bracket to get scanlMAfter' like functionality. See issue #677.

This can be further generalized to a type similar to Fold/Parser, giving it filtering and parsing capability as well (this is in fact equivalent to parseMany):

smapM :: (s -> a -> m (Step s b)) -> m s -> t m a -> t m b

take :: (IsStream t, Monad m) => Int -> t m a -> t m a

Take first n elements from the stream and discard the rest.

Since: 0.1.0

takeWhile :: (IsStream t, Monad m) => (a -> Bool) -> t m a -> t m a

End the stream as soon as the predicate fails on an element.

Since: 0.1.0

drop :: (IsStream t, Monad m) => Int -> t m a -> t m a

Discard first n elements from the stream and take the rest.

Since: 0.1.0

findIndices :: (IsStream t, Monad m) => (a -> Bool) -> t m a -> t m Int

Find all the indices where the element in the stream satisfies the given predicate.

findIndices = fold Fold.findIndices

Since: 0.5.0

intersperseM :: (IsStream t, MonadAsync m) => m a -> t m a -> t m a

Insert an effect and its output before consuming an element of a stream except the first one.

>>> S.toList $ S.trace putChar $ S.intersperseM (putChar . >> return ',') $ S.fromList "hello"
> h.,e.,l.,l.,o"h,e,l,l,o"

Since: 0.5.0

interjectSuffix :: (IsStream t, MonadAsync m) => Double -> m a -> t m a -> t m a

Intersperse a monadic action into the input stream after every n seconds.

> S.drain $ S.interjectSuffix 1 (putChar ',') $ S.mapM (\x -> threadDelay 1000000 >> putChar x) $ S.fromList "hello"
"h,e,l,l,o"

Internal

reverse :: (IsStream t, Monad m) => t m a -> t m a

Returns the elements of the stream in reverse order. The stream must be finite. Note that this necessarily buffers the entire stream in memory.

Since 0.7.0 (Monad m constraint)

Since: 0.1.1

reverse' :: (IsStream t, MonadIO m, Storable a) => t m a -> t m a

Like reverse but several times faster, requires a Storable instance.

Internal

Nesting

concatM :: (IsStream t, Monad m) => m (t m a) -> t m a

Given a stream value in the underlying monad, lift and join the underlying monad with the stream monad.

concatM = concat . yieldM
concatM = concat . lift    -- requires (MonadTrans t)
concatM = join . lift      -- requires (MonadTrans t, Monad (t m))

See also: concat, sequence

Internal

concatMapM :: (IsStream t, Monad m) => (a -> m (t m b)) -> t m a -> t m b

Map a stream producing monadic function on each element of the stream and then flatten the results into a single stream. Since the stream generation function is monadic, unlike concatMap, it can produce an effect at the beginning of each iteration of the inner loop.

Since: 0.6.0

splitOnSeq :: (IsStream t, MonadIO m, Storable a, Enum a, Eq a) => Array a -> Fold m a b -> t m a -> t m b

Like splitOn but the separator is a sequence of elements instead of a single element.

For illustration, let's define a function that operates on pure lists:

splitOnSeq' pat xs = S.toList $ S.splitOnSeq (A.fromList pat) (FL.toList) (S.fromList xs)

>>> splitOnSeq' "" "hello"
> ["h","e","l","l","o"]

>>> splitOnSeq' "hello" ""
> [""]

>>> splitOnSeq' "hello" "hello"
> ["",""]

>>> splitOnSeq' "x" "hello"
> ["hello"]

>>> splitOnSeq' "h" "hello"
> ["","ello"]

>>> splitOnSeq' "o" "hello"
> ["hell",""]

>>> splitOnSeq' "e" "hello"
> ["h","llo"]

>>> splitOnSeq' "l" "hello"
> ["he","","o"]

>>> splitOnSeq' "ll" "hello"
> ["he","o"]

splitOnSeq is an inverse of intercalate. The following law always holds:

intercalate . splitOn == id

The following law holds when the separator is non-empty and contains none of the elements present in the input lists:

splitOn . intercalate == id

Internal