Streamly.Internal.Data.Stream.IsStream.Nesting

Description

Transformations involving multiple streams, streams and unfolds or streams and folds.

Generate

Combining streams to generate streams.

Combine Two Streams

Functions ending in the shape:

t m a -> t m a -> t m a.

Appending

serial :: IsStream t => t m a -> t m a -> t m a infixr 6

Polymorphic version of the Semigroup operation <> of SerialT. Appends two streams sequentially, yielding all elements from the first stream, and then all elements from the second stream.

Since: 0.2.0 (Streamly)

Since: 0.8.0

append :: (IsStream t, Monad m) => t m b -> t m b -> t m b

Append the outputs of two streams, yielding all the elements from the first stream and then yielding all the elements from the second stream.

IMPORTANT NOTE: This could be 100x faster than serial/<> for appending a few (say 100) streams because it can fuse via stream fusion. However, it does not scale for a large number of streams (say 1000s) and becomes qudartically slow. Therefore use this for custom appending of a few streams but use concatMap or 'concatMapWith serial' for appending n streams or infinite containers of streams.

Internal

Interleaving

interleave :: (IsStream t, Monad m) => t m b -> t m b -> t m b

Interleaves the outputs of two streams, yielding elements from each stream alternately, starting from the first stream. If any of the streams finishes early the other stream continues alone until it too finishes.

>>> :set -XOverloadedStrings
>>> interleave "ab" ",,,," :: SerialT Identity Char
fromList "a,b,,,"
>>> interleave "abcd" ",," :: SerialT Identity Char
fromList "a,b,cd"

interleave is dual to interleaveMin, it can be called interleaveMax.

Do not use at scale in concatMapWith.

Internal

interleaveMin :: (IsStream t, Monad m) => t m b -> t m b -> t m b

Interleaves the outputs of two streams, yielding elements from each stream alternately, starting from the first stream. The output stops as soon as any of the two streams finishes, discarding the remaining part of the other stream. The last element of the resulting stream would be from the longer stream.

>>> :set -XOverloadedStrings
>>> interleaveMin "ab" ",,,," :: SerialT Identity Char
fromList "a,b,"
>>> interleaveMin "abcd" ",," :: SerialT Identity Char
fromList "a,b,c"

interleaveMin is dual to interleave.

Do not use at scale in concatMapWith.

Internal

interleaveSuffix :: (IsStream t, Monad m) => t m b -> t m b -> t m b

Interleaves the outputs of two streams, yielding elements from each stream alternately, starting from the first stream. As soon as the first stream finishes, the output stops, discarding the remaining part of the second stream. In this case, the last element in the resulting stream would be from the second stream. If the second stream finishes early then the first stream still continues to yield elements until it finishes.

>>> :set -XOverloadedStrings
>>> interleaveSuffix "abc" ",,,," :: SerialT Identity Char
fromList "a,b,c,"
>>> interleaveSuffix "abc" "," :: SerialT Identity Char
fromList "a,bc"

interleaveSuffix is a dual of interleaveInfix.

Do not use at scale in concatMapWith.

Internal

interleaveInfix :: (IsStream t, Monad m) => t m b -> t m b -> t m b

Interleaves the outputs of two streams, yielding elements from each stream alternately, starting from the first stream and ending at the first stream. If the second stream is longer than the first, elements from the second stream are infixed with elements from the first stream. If the first stream is longer then it continues yielding elements even after the second stream has finished.

>>> :set -XOverloadedStrings
>>> interleaveInfix "abc" ",,,," :: SerialT Identity Char
fromList "a,b,c"
>>> interleaveInfix "abc" "," :: SerialT Identity Char
fromList "a,bc"

interleaveInfix is a dual of interleaveSuffix.

Do not use at scale in concatMapWith.

Internal

wSerial :: IsStream t => t m a -> t m a -> t m a infixr 6

Polymorphic version of the Semigroup operation <> of WSerialT. Interleaves two streams, yielding one element from each stream alternately. When one stream stops the rest of the other stream is used in the output stream.

Since: 0.2.0 (Streamly)

Since: 0.8.0

wSerialFst :: IsStream t => t m a -> t m a -> t m a

Like wSerial but stops interleaving as soon as the first stream stops.

Since: 0.7.0

wSerialMin :: IsStream t => t m a -> t m a -> t m a

Like wSerial but stops interleaving as soon as any of the two streams stops.

Since: 0.7.0

Scheduling

ahead :: (IsStream t, MonadAsync m) => t m a -> t m a -> t m a infixr 6

Polymorphic version of the Semigroup operation <> of AheadT. Merges two streams sequentially but with concurrent lookahead.

Since: 0.3.0 (Streamly)

Since: 0.8.0

async :: (IsStream t, MonadAsync m) => t m a -> t m a -> t m a infixr 6

Polymorphic version of the Semigroup operation <> of AsyncT. Merges two streams possibly concurrently, preferring the elements from the left one when available.

Since: 0.2.0 (Streamly)

Since: 0.8.0

wAsync :: (IsStream t, MonadAsync m) => t m a -> t m a -> t m a infixr 6

Polymorphic version of the Semigroup operation <> of WAsyncT. Merges two streams concurrently choosing elements from both fairly.

Since: 0.2.0 (Streamly)

Since: 0.8.0

roundrobin :: (IsStream t, Monad m) => t m b -> t m b -> t m b

Schedule the execution of two streams in a fair round-robin manner, executing each stream once, alternately. Execution of a stream may not necessarily result in an output, a stream may chose to Skip producing an element until later giving the other stream a chance to run. Therefore, this combinator fairly interleaves the execution of two streams rather than fairly interleaving the output of the two streams. This can be useful in co-operative multitasking without using explicit threads. This can be used as an alternative to async.

Do not use at scale in concatMapWith.

Internal

Parallel

parallel :: (IsStream t, MonadAsync m) => t m a -> t m a -> t m a infixr 6

Execute two streams concurrently and merge their outputs. For example, if stream a is a serial stream consisting of a1, a2, a3 and stream b is a serial stream consisting of b1, b2, b3 then stream a parallel b may produce a1, b1, a2, b2, a3, b3 or a1, a2, b1, a3, b2, b3 or some other combination depending on the rate at which the two streams produce elements. However, the relative order of outputs from a single stream e.g. a1, a2, a3 would remain the same in the resulting stream. The effects in the two streams may occur concurrently.

To run single actions (instead of streams) in parallel wrap them into singleton streams. The following trivial example is semantically equivalent to running the action putStrLn "hello" in the current thread:

>>> S.toList $ S.yieldM (putStrLn "hello") `parallel` S.nil
> hello
> [()]

Run two actions concurrently:

>>> S.toList $ S.yieldM (putStrLn "hello") `parallel` S.yieldM (putStrLn "world")
> hello
> world
> [(),()]

Run effects concurrently, disregarding their outputs:

>>> S.toList $ S.nilM (putStrLn "hello") `parallel` S.nilM (putStrLn "world")
> hello
> world
> []

Run an effectful action, and a pure effect without any output, concurrently:

>>> S.toList $ S.yieldM (return 1) `parallel` S.nilM (putStrLn "world")
world
[1]

Note that parallel is a polymorphic version of the Semigroup operation <> of ParallelT.

nilM is currently Internal.

Since: 0.2.0 (Streamly)

Since: 0.8.0

parallelFst :: (IsStream t, MonadAsync m) => t m a -> t m a -> t m a

Like parallel but stops the output as soon as the first stream stops.

Internal

parallelMin :: (IsStream t, MonadAsync m) => t m a -> t m a -> t m a

Like parallel but stops the output as soon as any of the two streams stops.

Internal

Zipping

zipWith :: (IsStream t, Monad m) => (a -> b -> c) -> t m a -> t m b -> t m c

Zip two streams serially using a pure zipping function.

> S.toList $ S.zipWith (+) (S.fromList [1,2,3]) (S.fromList [4,5,6])
[5,7,9]

Since: 0.1.0

zipWithM :: (IsStream t, Monad m) => (a -> b -> m c) -> t m a -> t m b -> t m c

Like zipWith but using a monadic zipping function.

Since: 0.4.0

zipAsyncWith :: (IsStream t, MonadAsync m) => (a -> b -> c) -> t m a -> t m b -> t m c

Like zipWith but zips concurrently i.e. both the streams being zipped are generated concurrently.

Since: 0.1.0

zipAsyncWithM :: (IsStream t, MonadAsync m) => (a -> b -> m c) -> t m a -> t m b -> t m c

Like zipWithM but zips concurrently i.e. both the streams being zipped are generated concurrently.

Since: 0.4.0

Merging

mergeBy :: (IsStream t, Monad m) => (a -> a -> Ordering) -> t m a -> t m a -> t m a

Merge two streams using a comparison function. The head elements of both the streams are compared and the smaller of the two elements is emitted, if both elements are equal then the element from the first stream is used first.

If the streams are sorted in ascending order, the resulting stream would also remain sorted in ascending order.

> S.toList $ S.mergeBy compare (S.fromList [1,3,5]) (S.fromList [2,4,6,8])
[1,2,3,4,5,6,8]

Since: 0.6.0

mergeByM :: (IsStream t, Monad m) => (a -> a -> m Ordering) -> t m a -> t m a -> t m a

Like mergeBy but with a monadic comparison function.

Merge two streams randomly:

> randomly _ _ = randomIO >>= x -> return $ if x then LT else GT
> S.toList $ S.mergeByM randomly (S.fromList [1,1,1,1]) (S.fromList [2,2,2,2])
[2,1,2,2,2,1,1,1]

Merge two streams in a proportion of 2:1:

proportionately m n = do
 ref <- newIORef $ cycle $ concat [replicate m LT, replicate n GT]
 return $ \_ _ -> do
     r <- readIORef ref
     writeIORef ref $ tail r
     return $ head r

main = do
 f <- proportionately 2 1
 xs <- S.toList $ S.mergeByM f (S.fromList [1,1,1,1,1,1]) (S.fromList [2,2,2])
 print xs

[1,1,2,1,1,2,1,1,2]

Since: 0.6.0

mergeAsyncBy :: (IsStream t, MonadAsync m) => (a -> a -> Ordering) -> t m a -> t m a -> t m a

Like mergeBy but merges concurrently (i.e. both the elements being merged are generated concurrently).

Since: 0.6.0

mergeAsyncByM :: (IsStream t, MonadAsync m) => (a -> a -> m Ordering) -> t m a -> t m a -> t m a

Like mergeByM but merges concurrently (i.e. both the elements being merged are generated concurrently).

Since: 0.6.0

Trimming

dropPrefix :: t m a -> t m a -> t m a

Drop prefix from the input stream if present.

Space: O(1)

Unimplemented - Help wanted.

dropInfix :: t m a -> t m a -> t m a

Drop all matching infix from the input stream if present. Infix stream may be consumed multiple times.

Space: O(n) where n is the length of the infix.

Unimplemented - Help wanted.

dropSuffix :: t m a -> t m a -> t m a

Drop suffix from the input stream if present. Suffix stream may be consumed multiple times.

Space: O(n) where n is the length of the suffix.

Unimplemented - Help wanted.

Combine N Streams

Functions generally ending in these shapes:

concat: f (t m a) -> t m a
concatMap: (a -> t m b) -> t m a -> t m b
concatUnfold: Unfold m a b -> t m a -> t m b

Flatten Containers of Streams

concatM :: (IsStream t, Monad m) => m (t m a) -> t m a

Given a stream value in the underlying monad, lift and join the underlying monad with the stream monad.

concatM = concat . yieldM
concatM = concat . lift    -- requires (MonadTrans t)
concatM = join . lift      -- requires (MonadTrans t, Monad (t m))

See also: concat, sequence

Internal

concat :: (IsStream t, Monad m) => t m (t m a) -> t m a

Flatten a stream of streams to a single stream.

concat = concatMap id

Internal

concatFoldableWith :: (IsStream t, Foldable f) => (t m a -> t m a -> t m a) -> f (t m a) -> t m a

A variant of fold that allows you to fold a Foldable container of streams using the specified stream sum operation.

concatFoldableWith async $ map return [1..3]

Equivalent to:

concatFoldableWith f = S.concatMapFoldableWith f id

Since: 0.8.0 (Renamed foldWith to concatFoldableWith)

Since: 0.1.0 (Streamly)

ConcatMap

Map and flatten streams.

concatMapFoldableWith :: (IsStream t, Foldable f) => (t m b -> t m b -> t m b) -> (a -> t m b) -> f a -> t m b

A variant of foldMap that allows you to map a monadic streaming action on a Foldable container and then fold it using the specified stream merge operation.

concatMapFoldableWith async return [1..3]

Equivalent to:

concatMapFoldableWith f g xs = S.concatMapWith f g (S.fromFoldable xs)

Since: 0.8.0 (Renamed foldMapWith to concatMapFoldableWith)

Since: 0.1.0 (Streamly)

concatForFoldableWith :: (IsStream t, Foldable f) => (t m b -> t m b -> t m b) -> f a -> (a -> t m b) -> t m b

Like concatMapFoldableWith but with the last two arguments reversed i.e. the monadic streaming function is the last argument.

Equivalent to:

concatForFoldableWith = flip S.concatMapFoldableWith

Since: 0.8.0 (Renamed forEachWith to concatForFoldableWith)

Since: 0.1.0 (Streamly)

concatMap :: (IsStream t, Monad m) => (a -> t m b) -> t m a -> t m b

Map a stream producing function on each element of the stream and then flatten the results into a single stream.

concatMap f = 'concat . map f'
concatMap = concatMapWith serial
concatMap f = concatMapM (return . f)
concatMap f = concatUnfold (UF.lmap f UF.fromStream)

Since: 0.6.0

concatMapM :: (IsStream t, Monad m) => (a -> m (t m b)) -> t m a -> t m b

Map a stream producing monadic function on each element of the stream and then flatten the results into a single stream. Since the stream generation function is monadic, unlike concatMap, it can produce an effect at the beginning of each iteration of the inner loop.

Since: 0.6.0

concatMapWith :: IsStream t => (t m b -> t m b -> t m b) -> (a -> t m b) -> t m a -> t m b

concatMapWith merge map stream is a two dimensional looping combinator. The first argument specifies a merge or concat function that is used to merge the streams generated by applying the second argument i.e. the map function to each element of the input stream. The concat function could be serial, parallel, async, ahead or any other zip or merge function and the second argument could be any stream generation function using a seed.

Compare foldMapWith

Since: 0.7.0

Since: 0.8.0 (signature change)

concatSmapMWith :: (IsStream t, Monad m) => (t m b -> t m b -> t m b) -> (s -> a -> m (s, t m b)) -> m s -> t m a -> t m b

Like concatMapWith but carries a state which can be used to share information across multiple steps of concat.

concatSmapMWith combine f initial = concatMapWith combine id . smapM f initial

Internal

ConcatUnfold

Unfold and flatten streams.

concatUnfold :: (IsStream t, Monad m) => Unfold m a b -> t m a -> t m b

Like concatMap but uses an Unfold for stream generation. Unlike concatMap this can fuse the Unfold code with the inner loop and therefore provide many times better performance.

Since: 0.7.0

concatUnfoldInterleave :: (IsStream t, Monad m) => Unfold m a b -> t m a -> t m b

Like concatUnfold but interleaves the streams in the same way as interleave behaves instead of appending them.

Internal

concatUnfoldRoundrobin :: (IsStream t, Monad m) => Unfold m a b -> t m a -> t m b

Like concatUnfold but executes the streams in the same way as roundrobin.

Internal

Interpose

Insert effects between streams. Like concatUnfold but intersperses an effect between the streams. A special case of gintercalate.

interpose :: (IsStream t, Monad m) => c -> Unfold m b c -> t m b -> t m c

Unfold the elements of a stream, intersperse the given element between the unfolded streams and then concat them into a single stream.

unwords = S.interpose ' '

Internal

interposeSuffix :: (IsStream t, Monad m) => c -> Unfold m b c -> t m b -> t m c

Unfold the elements of a stream, append the given element after each unfolded stream and then concat them into a single stream.

unlines = S.interposeSuffix '\n'

Internal

Intercalate

Insert Streams between Streams. Like concatUnfold but intersperses streams from another source between the streams from the first source.

intercalate :: (IsStream t, Monad m) => b -> Unfold m b c -> t m b -> t m c

intersperse followed by unfold and concat.

unwords = intercalate " " UF.fromList

>>> intercalate " " UF.fromList ["abc", "def", "ghi"]
> "abc def ghi"

Internal

intercalateSuffix :: (IsStream t, Monad m) => b -> Unfold m b c -> t m b -> t m c

intersperseSuffix followed by unfold and concat.

unlines = intercalateSuffix "\n" UF.fromList

>>> intercalate "\n" UF.fromList ["abc", "def", "ghi"]
> "abc\ndef\nghi\n"

Internal

gintercalate :: (IsStream t, Monad m) => Unfold m a c -> t m a -> Unfold m b c -> t m b -> t m c

interleaveInfix followed by unfold and concat.

Internal

gintercalateSuffix :: (IsStream t, Monad m) => Unfold m a c -> t m a -> Unfold m b c -> t m b -> t m c

interleaveSuffix followed by unfold and concat.

Internal

IterateMap

Map and flatten Trees of Streams

iterateMapWith :: IsStream t => (t m a -> t m a -> t m a) -> (a -> t m a) -> t m a -> t m a

Like iterateM but iterates after mapping a stream generator on the output.

Yield an input element in the output stream, map a stream generator on it and then do the same on the resulting stream. This can be used for a depth first traversal of a tree like structure.

Note that iterateM is a special case of iterateMapWith:

iterateM f = iterateMapWith serial (yieldM . f) . yieldM

It can be used to traverse a tree structure. For example, to list a directory tree:

Stream.iterateMapWith Stream.serial
    (either Dir.toEither (const nil))
    (yield (Left "tmp"))

Internal

iterateSmapMWith :: (IsStream t, Monad m) => (t m a -> t m a -> t m a) -> (b -> a -> m (b, t m a)) -> m b -> t m a -> t m a

Like iterateMap but carries a state in the stream generation function. This can be used to traverse graph like structures, we can remember the visited nodes in the state to avoid cycles.

Note that a combination of iterateMap and usingState can also be used to traverse graphs. However, this function provides a more localized state instead of using a global state.

See also: mfix

Internal

iterateMapLeftsWith :: IsStream t => (t m (Either a b) -> t m (Either a b) -> t m (Either a b)) -> (a -> t m (Either a b)) -> t m (Either a b) -> t m (Either a b)

In an Either stream iterate on Lefts. This is a special case of iterateMapWith:

iterateMapLeftsWith combine f = iterateMapWith combine (either f (const nil))

To traverse a directory tree:

iterateMapLeftsWith serial Dir.toEither (yield (Left "tmp"))

Internal

Eliminate

Folding and Parsing chunks of streams to eliminate nested streams. Functions generally ending in these shapes:

f (Fold m a b) -> t m a -> t m b
f (Parser m a b) -> t m a -> t m b

Folding

Apply folds on a stream.

foldMany :: (IsStream t, Monad m) => Fold m a b -> t m a -> t m b

Apply a Fold repeatedly on a stream and emit the parsed values in the output stream.

This is the streaming dual of the many parse combinator.

>>> f = Fold.takeLE 2 Fold.sum
>>> Stream.toList $ Stream.foldMany f $ Stream.fromList [1..10]
> [3,7,11,15,19]

>>> f = Fold.sliceEndWith Fold.toList
>>> Stream.toList $ Stream.foldMany f $ Stream.fromList "hello\nworld"
> ["hello\n","world"]

Internal

foldSequence :: t m (Fold m a b) -> t m a -> t m b

Apply a stream of folds to an input stream and emit the results in the output stream.

Internal

foldIterate :: (b -> Fold m a b) -> b -> t m a -> t m b

Iterate a fold generator on a stream. The initial value b is used to generate the first fold, the fold is applied on the stream and the result of the fold is used to generate the next fold and so on.

>>> f x = Fold.takeLE 2 (Fold.mconcatTo x)
>>> s = Stream.map Sum $ Stream.fromList [1..10]
>>> Stream.toList $ Stream.map getSum $ Stream.foldIterate f 0 s
> [3,10,21,36,55,55]

This is the streaming equivalent of monad like sequenced application of folds where next fold is dependent on the previous fold.

Internal

Parsing

Apply parsers on a stream.

parseMany :: (IsStream t, MonadThrow m) => Parser m a b -> t m a -> t m b

Apply a Parser repeatedly on a stream and emit the parsed values in the output stream.

This is the streaming equivalent of the many parse combinator.

>>> S.toList $ S.parseMany (PR.take 2 $ PR.fromFold FL.sum) $ S.fromList [1..10]
> [3,7,11,15,19]

>>> S.toList $ S.parseMany (PR.line FL.toList) $ S.fromList "hello\nworld"
> ["hello\n","world"]

Internal

parseManyD :: (IsStream t, MonadThrow m) => Parser m a b -> t m a -> t m b

parseManyTill :: Parser m a b -> Parser m a x -> t m a -> t m b

parseManyTill collect test stream tries the parser test on the input, if test fails it backtracks and tries collect, after collect succeeds test is tried again and so on. The parser stops when test succeeds. The output of test is discarded and the output of collect is emitted in the output stream. The parser fails if collect fails.

Unimplemented

parseSequence :: t m (Parser m a b) -> t m a -> t m b

Apply a stream of parsers to an input stream and emit the results in the output stream.

Internal

parseIterate :: (IsStream t, MonadThrow m) => (b -> Parser m a b) -> b -> t m a -> t m b

Iterate a parser generating function on a stream. The initial value b is used to generate the first parser, the parser is applied on the stream and the result is used to generate the next parser and so on.

>>> S.toList $ S.map getSum $ S.parseIterate (\b -> PR.take 2 (FL.mconcatTo b)) 0 $ S.map Sum $ S.fromList [1..10]
> [3,10,21,36,55,55]

This is the streaming equivalent of monad like sequenced application of parsers where next parser is dependent on the previous parser.

Internal

Chunking

Element unaware grouping.

chunksOf :: (IsStream t, Monad m) => Int -> Fold m a b -> t m a -> t m b

Group the input stream into groups of n elements each and then fold each group using the provided fold function.

> S.toList $ S.chunksOf 2 FL.sum (S.enumerateFromTo 1 10)
 [3,7,11,15,19]

This can be considered as an n-fold version of takeLE where we apply takeLE repeatedly on the leftover stream until the stream exhausts.

chunksOf n f = foldMany (FL.takeLE n f)

Since: 0.7.0

chunksOf2 :: (IsStream t, Monad m) => Int -> m c -> Fold2 m c a b -> t m a -> t m b

Internal

arraysOf :: (IsStream t, MonadIO m, Storable a) => Int -> t m a -> t m (Array a)

arraysOf n stream groups the elements in the input stream into arrays of n elements each.

Same as the following but may be more efficient:

arraysOf n = Stream.foldMany (A.writeN n)

Internal

intervalsOf :: (IsStream t, MonadAsync m) => Double -> Fold m a b -> t m a -> t m b

Group the input stream into windows of n second each and then fold each group using the provided fold function.

Since: 0.7.0

Grouping

groups :: (IsStream t, Monad m, Eq a) => Fold m a b -> t m a -> t m b

groups = groupsBy (==)
groups = groupsByRolling (==)

Groups contiguous spans of equal elements together in individual groups.

>>> S.toList $ S.groups FL.toList $ S.fromList [1,1,2,2]
> [[1,1],[2,2]]

Since: 0.7.0

groupsBy :: (IsStream t, Monad m) => (a -> a -> Bool) -> Fold m a b -> t m a -> t m b

groupsBy cmp f $ S.fromList [a,b,c,...] assigns the element a to the first group, if b `cmp` a is True then b is also assigned to the same group. If c `cmp` a is True then c is also assigned to the same group and so on. When the comparison fails a new group is started. Each group is folded using the fold f and the result of the fold is emitted in the output stream.

>>> S.toList $ S.groupsBy (>) FL.toList $ S.fromList [1,3,7,0,2,5]
> [[1,3,7],[0,2,5]]

Since: 0.7.0

groupsByRolling :: (IsStream t, Monad m) => (a -> a -> Bool) -> Fold m a b -> t m a -> t m b

Unlike groupsBy this function performs a rolling comparison of two successive elements in the input stream. groupsByRolling cmp f $ S.fromList [a,b,c,...] assigns the element a to the first group, if a `cmp` b is True then b is also assigned to the same group. If b `cmp` c is True then c is also assigned to the same group and so on. When the comparison fails a new group is started. Each group is folded using the fold f.

>>> S.toList $ S.groupsByRolling (\a b -> a + 1 == b) FL.toList $ S.fromList [1,2,3,7,8,9]
> [[1,2,3],[7,8,9]]

Since: 0.7.0

Splitting

Streams can be sliced into segments in space or in time. We use the term chunk to refer to a spatial length of the stream (spatial window) and the term session to refer to a length in time (time window).

splitOn :: (IsStream t, Monad m) => (a -> Bool) -> Fold m a b -> t m a -> t m b

Split on an infixed separator element, dropping the separator. Splits the stream on separator elements determined by the supplied predicate, separator is considered as infixed between two segments, if one side of the separator is missing then it is parsed as an empty stream. The supplied Fold is applied on the split segments. With - representing non-separator elements and . as separator, splitOn splits as follows:

"--.--" => "--" "--"
"--."   => "--" ""
".--"   => ""   "--"

splitOn (== x) is an inverse of intercalate (S.yield x)

Let's use the following definition for illustration:

splitOn' p xs = S.toList $ S.splitOn p (FL.toList) (S.fromList xs)

>>> splitOn' (== '.') ""
[""]

>>> splitOn' (== '.') "."
["",""]

>>> splitOn' (== '.') ".a"
> ["","a"]

>>> splitOn' (== '.') "a."
> ["a",""]

>>> splitOn' (== '.') "a.b"
> ["a","b"]

>>> splitOn' (== '.') "a..b"
> ["a","","b"]

Since: 0.7.0

splitOnSuffix :: (IsStream t, Monad m) => (a -> Bool) -> Fold m a b -> t m a -> t m b

Like splitOn but the separator is considered as suffixed to the segments in the stream. A missing suffix at the end is allowed. A separator at the beginning is parsed as empty segment. With - representing elements and . as separator, splitOnSuffix splits as follows:

 "--.--." => "--" "--"
 "--.--"  => "--" "--"
 ".--."   => "" "--"

splitOnSuffix' p xs = S.toList $ S.splitSuffixBy p (FL.toList) (S.fromList xs)

>>> splitOnSuffix' (== '.') ""
[]

>>> splitOnSuffix' (== '.') "."
[""]

>>> splitOnSuffix' (== '.') "a"
["a"]

>>> splitOnSuffix' (== '.') ".a"
> ["","a"]

>>> splitOnSuffix' (== '.') "a."
> ["a"]

>>> splitOnSuffix' (== '.') "a.b"
> ["a","b"]

>>> splitOnSuffix' (== '.') "a.b."
> ["a","b"]

>>> splitOnSuffix' (== '.') "a..b.."
> ["a","","b",""]

lines = splitOnSuffix (== '\n')

Since: 0.7.0

splitWithSuffix :: (IsStream t, Monad m) => (a -> Bool) -> Fold m a b -> t m a -> t m b

Like splitOnSuffix but keeps the suffix attached to the resulting splits.

splitWithSuffix' p xs = S.toList $ S.splitWithSuffix p (FL.toList) (S.fromList xs)

>>> splitWithSuffix' (== '.') ""
[]

>>> splitWithSuffix' (== '.') "."
["."]

>>> splitWithSuffix' (== '.') "a"
["a"]

>>> splitWithSuffix' (== '.') ".a"
> [".","a"]

>>> splitWithSuffix' (== '.') "a."
> ["a."]

>>> splitWithSuffix' (== '.') "a.b"
> ["a.","b"]

>>> splitWithSuffix' (== '.') "a.b."
> ["a.","b."]

>>> splitWithSuffix' (== '.') "a..b.."
> ["a.",".","b.","."]

Since: 0.7.0

wordsBy :: (IsStream t, Monad m) => (a -> Bool) -> Fold m a b -> t m a -> t m b

Like splitOn after stripping leading, trailing, and repeated separators. Therefore, ".a..b." with . as the separator would be parsed as ["a","b"]. In other words, its like parsing words from whitespace separated text.

wordsBy' p xs = S.toList $ S.wordsBy p (FL.toList) (S.fromList xs)

>>> wordsBy' (== ',') ""
> []

>>> wordsBy' (== ',') ","
> []

>>> wordsBy' (== ',') ",a,,b,"
> ["a","b"]

words = wordsBy isSpace

Since: 0.7.0

splitBySeq :: (IsStream t, MonadAsync m, Storable a, Enum a, Eq a) => Array a -> Fold m a b -> t m a -> t m b

Like splitOnSeq but splits the separator as well, as an infix token.

splitOn'_ pat xs = S.toList $ S.splitOn' (A.fromList pat) (FL.toList) (S.fromList xs)

>>> splitOn'_ "" "hello"
> ["h","","e","","l","","l","","o"]

>>> splitOn'_ "hello" ""
> [""]

>>> splitOn'_ "hello" "hello"
> ["","hello",""]

>>> splitOn'_ "x" "hello"
> ["hello"]

>>> splitOn'_ "h" "hello"
> ["","h","ello"]

>>> splitOn'_ "o" "hello"
> ["hell","o",""]

>>> splitOn'_ "e" "hello"
> ["h","e","llo"]

>>> splitOn'_ "l" "hello"
> ["he","l","","l","o"]

>>> splitOn'_ "ll" "hello"
> ["he","ll","o"]

Internal

splitOnSeq :: (IsStream t, MonadIO m, Storable a, Enum a, Eq a) => Array a -> Fold m a b -> t m a -> t m b

Like splitOn but the separator is a sequence of elements instead of a single element.

For illustration, let's define a function that operates on pure lists:

splitOnSeq' pat xs = S.toList $ S.splitOnSeq (A.fromList pat) (FL.toList) (S.fromList xs)

>>> splitOnSeq' "" "hello"
> ["h","e","l","l","o"]

>>> splitOnSeq' "hello" ""
> [""]

>>> splitOnSeq' "hello" "hello"
> ["",""]

>>> splitOnSeq' "x" "hello"
> ["hello"]

>>> splitOnSeq' "h" "hello"
> ["","ello"]

>>> splitOnSeq' "o" "hello"
> ["hell",""]

>>> splitOnSeq' "e" "hello"
> ["h","llo"]

>>> splitOnSeq' "l" "hello"
> ["he","","o"]

>>> splitOnSeq' "ll" "hello"
> ["he","o"]

splitOnSeq is an inverse of intercalate. The following law always holds:

intercalate . splitOn == id

The following law holds when the separator is non-empty and contains none of the elements present in the input lists:

splitOn . intercalate == id

Internal

splitOnSuffixSeq :: (IsStream t, MonadIO m, Storable a, Enum a, Eq a) => Array a -> Fold m a b -> t m a -> t m b

Like splitSuffixBy but the separator is a sequence of elements, instead of a predicate for a single element.

splitOnSuffixSeq_ pat xs = S.toList $ S.splitOnSuffixSeq (A.fromList pat) (FL.toList) (S.fromList xs)

>>> splitOnSuffixSeq_ "." ""
[]

>>> splitOnSuffixSeq_ "." "."
[""]

>>> splitOnSuffixSeq_ "." "a"
["a"]

>>> splitOnSuffixSeq_ "." ".a"
> ["","a"]

>>> splitOnSuffixSeq_ "." "a."
> ["a"]

>>> splitOnSuffixSeq_ "." "a.b"
> ["a","b"]

>>> splitOnSuffixSeq_ "." "a.b."
> ["a","b"]

>>> splitOnSuffixSeq_ "." "a..b.."
> ["a","","b",""]

lines = splitOnSuffixSeq "\n"

splitOnSuffixSeq is an inverse of intercalateSuffix. The following law always holds:

intercalateSuffix . splitOnSuffixSeq == id

The following law holds when the separator is non-empty and contains none of the elements present in the input lists:

splitSuffixOn . intercalateSuffix == id

Internal

splitWithSuffixSeq :: (IsStream t, MonadIO m, Storable a, Enum a, Eq a) => Array a -> Fold m a b -> t m a -> t m b

Like splitOnSuffixSeq but keeps the suffix intact in the splits.

splitWithSuffixSeq'_ pat xs = S.toList $ S.splitWithSuffixSeq (A.fromList pat) (FL.toList) (S.fromList xs)

>>> splitWithSuffixSeq' "." ""
[""]

>>> splitWithSuffixSeq' "." "."
["."]

>>> splitWithSuffixSeq' "." "a"
["a"]

>>> splitWithSuffixSeq' "." ".a"
> [".","a"]

>>> splitWithSuffixSeq' "." "a."
> ["a."]

>>> splitWithSuffixSeq' "." "a.b"
> ["a.","b"]

>>> splitWithSuffixSeq' "." "a.b."
> ["a.","b."]

>>> splitWithSuffixSeq' "." "a..b.."
> ["a.",".","b.","."]

Internal

Windowed Classification

Split the stream into windows or chunks in space or time. Each window can be associated with a key, all events associated with a particular key in the window can be folded to a single result. The stream is split into windows of specified size, the window can be terminated early if the closing flag is specified in the input stream.

The term "chunk" is used for a space window and the term "session" is used for a time window.

Tumbling Windows

A new window starts after the previous window is finished.

classifySessionsBy

Arguments

:: (IsStream t, MonadAsync m, Ord k)

=> Double
timer tick in seconds

-> Double
session timeout in seconds

-> Bool
reset the timeout when an event is received

-> (Int -> m Bool)
predicate to eject sessions based on session count

-> Fold m a b
Fold to be applied to session events

-> t m (k, a, AbsTime)
session key, data, timestamp

-> t m (k, b)
session key, fold result

classifySessionsBy tick timeout idle pred f stream groups timestamped events in an input event stream into sessions based on a session key. Each element in the input stream is an event consisting of a triple (session key, sesssion data, timestamp). session key is a key that uniquely identifies the session. All the events belonging to a session are folded using the fold f until the fold terminates or a timeout has occurred. The session key and the result of the fold are emitted in the output stream when the session is purged.

When idle is False, timeout is the maximum lifetime of a session in seconds, measured from the timestamp of the first event in that session. When idle is True then the timeout is an idle timeout, it is reset after every event received in the session.

timestamp in an event characterizes the time when the input event was generated, this is an absolute time measured from some Epoch. The notion of current time is maintained by a monotonic event time clock using the timestamps seen in the input stream. The latest timestamp seen till now is used as the base for the current time. When no new events are seen, a timer is started with a tick duration specified by tick. This timer is used to detect session timeouts in the absence of new events.

The predicate pred is invoked with the current session count, if it returns True a session is ejected from the session cache before inserting a new session. This could be useful to alert or eject sessions when the number of sessions becomes too high.

Internal

classifySessionsOf

Arguments

:: (IsStream t, MonadAsync m, Ord k)

=> Double
time window size

-> (Int -> m Bool)
predicate to eject sessions on session count

-> Fold m a b
Fold to be applied to session events

-> t m (k, a, AbsTime)
session key, data, timestamp

-> t m (k, b)

Split the stream into fixed size time windows of specified interval in seconds. Within each such window, fold the elements in sessions identified by the session keys. The fold result is emitted in the output stream if the fold returns a Left result or if the time window ends.

Session timestamp in the input stream is an absolute time from some epoch, characterizing the time when the input element was generated. To detect session window end, a monotonic event time clock is maintained synced with the timestamps with a clock resolution of 1 second.

If the ejection predicate returns True, the session with the longest lifetime is ejected before inserting a new session.

classifySessionsOf interval pred = classifySessionsBy 1 interval False pred

>>> S.mapM_ print
  $ S.classifySessionsOf 3 (const (return False)) (fmap Right FL.toList)
  $ S.map ((ts,(k,a)) -> (k, a, ts))
  $ S.timestamped
  $ S.delay 1
  $ (,) $ S.fromList [1,2,3] * S.fromList [1,2,3]

Internal

Keep Alive Windows

The window size is extended if an event arrives within the specified window size. This can represent sessions with idle or inactive timeout.

classifyKeepAliveSessions

Arguments

:: (IsStream t, MonadAsync m, Ord k)

=> Double
session inactive timeout

-> (Int -> m Bool)
predicate to eject sessions on session count

-> Fold m a b
Fold to be applied to session payload data

-> t m (k, a, AbsTime)
session key, data, timestamp

-> t m (k, b)

Like classifySessionsOf but the session is kept alive if an event is received within the session window. The session times out and gets closed only if no event is received within the specified session window size.

If the ejection predicate returns True, the session that was idle for the longest time is ejected before inserting a new session.

classifyKeepAliveSessions timeout pred = classifySessionsBy 1 timeout True pred

Internal

Transform (Nested Containers)

splitInnerBy :: (IsStream t, Monad m) => (f a -> m (f a, Maybe (f a))) -> (f a -> f a -> m (f a)) -> t m (f a) -> t m (f a)

splitInnerBy splitter joiner stream splits the inner containers f a of an input stream t m (f a) using the splitter function. Container elements f a are collected until a split occurs, then all the elements before the split are joined using the joiner function.

For example, if we have a stream of Array Word8, we may want to split the stream into arrays representing lines separated by 'n' byte such that the resulting stream after a split would be one array for each line.

CAUTION! This is not a true streaming function as the container size after the split and merge may not be bounded.

Internal

splitInnerBySuffix :: (IsStream t, Monad m, Eq (f a), Monoid (f a)) => (f a -> m (f a, Maybe (f a))) -> (f a -> f a -> m (f a)) -> t m (f a) -> t m (f a)

Like splitInnerBy but splits assuming the separator joins the segment in a suffix style.

Internal