Streamly.Internal.Data.Stream.StreamD

Description

Direct style re-implementation of CPS stream in Streamly.Internal.Data.Stream.StreamK. The symbol or suffix D in this module denotes the Direct style. GHC is able to INLINE and fuse direct style better, providing better performance than CPS implementation.

import qualified Streamly.Internal.Data.Stream.StreamD as D

Documentation

module Streamly.Internal.Data.Stream.StreamD.Type

module Streamly.Internal.Data.Stream.StreamD.Generate

module Streamly.Internal.Data.Stream.StreamD.Eliminate

module Streamly.Internal.Data.Stream.StreamD.Exception

module Streamly.Internal.Data.Stream.StreamD.Lift

module Streamly.Internal.Data.Stream.StreamD.Nesting

module Streamly.Internal.Data.Stream.StreamD.Transform