Streamly.Internal.Data.Time.Units

Description

Time Unit Conversions

class TimeUnit a

A type class for converting between time units using Integer as the intermediate and the widest representation with a nanosecond resolution. This system of units can represent arbitrarily large times but provides least efficient arithmetic operations due to Integer arithmetic.

NOTE: Converting to and from units may truncate the value depending on the original value and the size and resolution of the destination unit.

A type class for converting between units of time using TimeSpec as the intermediate representation. This system of units can represent up to ~292 billion years at nanosecond resolution with reasonably efficient arithmetic operations.

NOTE: Converting to and from units may truncate the value depending on the original value and the size and resolution of the destination unit.

Minimal complete definition

toTimeSpec, fromTimeSpec

class TimeUnit64 a

A type class for converting between units of time using Int64 as the intermediate representation with a nanosecond resolution. This system of units can represent up to ~292 years at nanosecond resolution with fast arithmetic operations.

NOTE: Converting to and from units may truncate the value depending on the original value and the size and resolution of the destination unit.

Minimal complete definition

toNanoSecond64, fromNanoSecond64

Time Units

data TimeSpec

Data type to represent practically large quantities of time efficiently. It can represent time up to ~292 billion years at nanosecond resolution.

Constructors

TimeSpec

Fields

sec :: !Int64

seconds

nsec :: !Int64

nanoseconds

newtype NanoSecond64

An Int64 time representation with a nanosecond resolution. It can represent time up to ~292 years.

Constructors

NanoSecond64 Int64

newtype MicroSecond64

An Int64 time representation with a microsecond resolution. It can represent time up to ~292,000 years.

Constructors

MicroSecond64 Int64

newtype MilliSecond64

An Int64 time representation with a millisecond resolution. It can represent time up to ~292 million years.

Constructors

MilliSecond64 Int64

showNanoSecond64 :: NanoSecond64 -> String

Convert nanoseconds to a string showing time in an appropriate unit.

Absolute times (using TimeSpec)

newtype AbsTime

Absolute times are relative to a predefined epoch in time. AbsTime represents times using TimeSpec which can represent times up to ~292 billion years at a nanosecond resolution.

Constructors

AbsTime TimeSpec

toAbsTime :: TimeUnit a => a -> AbsTime

Convert a TimeUnit to an absolute time.

fromAbsTime :: TimeUnit a => AbsTime -> a

Convert absolute time to a TimeUnit.

Relative times (using TimeSpec)

data RelTime

toRelTime :: TimeUnit a => a -> RelTime

fromRelTime :: TimeUnit a => RelTime -> a

diffAbsTime :: AbsTime -> AbsTime -> RelTime

addToAbsTime :: AbsTime -> RelTime -> AbsTime

Relative times (using NanoSecond64)

data RelTime64

Relative times are relative to some arbitrary point of time. Unlike AbsTime they are not relative to a predefined epoch.

toRelTime64 :: TimeUnit64 a => a -> RelTime64

Convert a TimeUnit to a relative time.

fromRelTime64 :: TimeUnit64 a => RelTime64 -> a

Convert relative time to a TimeUnit.

diffAbsTime64 :: AbsTime -> AbsTime -> RelTime64

Difference between two absolute points of time.

addToAbsTime64 :: AbsTime -> RelTime64 -> AbsTime

showRelTime64 :: RelTime64 -> String