Streamly.Internal.Data.Unfold.Types

Description

Documentation

data Unfold m a b

An Unfold m a b is a generator of a stream of values of type b from a seed of type a in Monad m.

Since: 0.7.0

Constructors

Unfold (s -> m (Step s b)) (a -> m s)
Unfold step inject