Streamly.Internal.Data.Zipper

Description

A special zipper derived from a stream to facilitate stream parsing.

Zipper

The zipper is designed to enable moving backward and forward in the stream efficiently to be able to implement backtracking parsers. The zipper can buffer data until the parser indicates that it is not needed, we can move back in the buffered data if the parser needs to backtrack.

Checkpointing

We have a checkpointing mechanism built into the zipper to handle backtracking. Whenever we enter a computation that can fail and has an alternative to run then we add a checkpoint to the zipper. The checkpoint ensures that we do not release the buffer upto the checkpoint until this computation succeeds, if the computation fails we can backtrack to focus the cursor back to the checkpoint position.

Future Work

We can possibly have a typeclass providing a zipper interface so that we can derive a zipper from different data sources efficiently. For example,

stream zipper

zippers from different types of arrays

zipper from a tree of arrays for dynamic in-memory storage

file handle zipper can utilize the seek to go back and forth in the file

A tree of zippers where a lower level zipper is dependent on (references data in) a higher level zipper could be useful in implementing distributive parsers.

See https://hackage.haskell.org/package/zipper for an existing type class. We need a monadic one.

Documentation

data Zipper m a

Zipper checkpoints lefts rights tail. The focus is on the first element of rights. lefts is buffered data on the right of the cursor. Note that lefts is stored as a reversed list, this helps is adding more items to the list quickly. tail is a stream that is used to generate more data if the cursor moves past rights.

checkpoints is a stack of checkpoints. A new checkpoint is created by a checkpoint operation. A checkpoint consists of a count that tracks how many elements we have yielded after the checkpoint was created. We need this information to backtrack to the checkpoint. If we enter a nested alternative we add another checkpoint in the stack. When we exit an alternative we call a release on the checkpoint. The release removes the checkpoint from the stack and adds its element count to the previous checkpoint in the stack. When the last checkpoint is removed the buffer is released.

Internal

Constructors

Zipper [Int] [a] [a] (Stream m a)

Construction

nil :: Zipper m a

Empty zipper.

Internal

fromStream :: Stream m a -> Zipper m a

Create a zipper from a stream.

Internal

fromList :: [a] -> Zipper m a

Create a zipper from a list.

Internal

Checkpointing

checkpoint :: Zipper m a -> Zipper m a

Add a checkpoint to the Zipper so that we do not release the buffer beyond the checkpoint.

Internal

release :: Zipper m a -> Zipper m a

Release the latest checkpoint, releases any values held by the checkpoint. Note that the values may still be held by other checkpoints in the stack of checkpoints.

Internal

restore :: Zipper m a -> Zipper m a

Rewind to restore the cursor to the latest checkpoint.

Internal

Parsing

parse :: MonadCatch m => (s -> a -> m (Step s b)) -> m s -> (s -> m b) -> Zipper m a -> m (Zipper m a, Either String b)

Parse a stream zipper using a parser's step, initial and extract functions.