Streamly.Internal.FileSystem.Event.Linux

Description

Overview

Use watchPaths with a list of file system paths you want to watch as argument. It returns a stream of Event representing the file system events occurring under the watched paths.

Stream.mapM_ (putStrLn . showEvent) $ watchPaths [Array.fromCString# "dir"#]

Event is an opaque type. Accessor functions (e.g. showEvent above) provided in this module are used to determine the attributes of the event.

Identical successive events may be coalesced into a single event.

Design notes

For reference documentation see:

inotify man page

We try to keep the macOS/Linux/Windows event handling APIs and defaults semantically and syntactically as close as possible.

BUGs

When testing on Linux Kernel version 5.3.0-53-generic #47-Ubuntu, the last event for the root path seems to be delayed until one more event occurs.

Differences between macOS and Linux APIs:

macOS watch is based on the path provided to it, if the path is deleted and recreated it will still be watched, if the path moves to another path it won't be watched anymore. Whereas Linux watch is based on a handle to the path, if the path is deleted and recreated it won't be watched, if the path moves to another it can still be watched (though this is configurable).

macOS watches the directory hierarchy recursively, Linux watches only one level of dir, recursive watch has to be built in user space by watching for create events and adding the new directories to the watch. Not sure how this will scale for too many paths.

In macOS the path of the subject of the event is absolute, in Linux the path is the name of the object inside the dir being watched.

On Linux watchPaths fails if a path does not exist, on macOS it does not fail.

Subscribing to events

Default configuration

data Config

Watch configuration, used to specify the events of interest and the behavior of the watch.

Internal

data Toggle

Whether a setting is On or Off.

Internal

Constructors

On

Off

defaultConfig :: Config

The default is:

setFollowSymLinks On

setUnwatchMoved On

setOneShot Off

setOnlyDir Off

setWhenExists AddIfExists

setAllEvents On

Internal

Watch Behavior

setFollowSymLinks :: Toggle -> Config -> Config

If the pathname to be watched is a symbolic link then watch the target of the symbolic link instead of the symbolic link itself.

default: On

Internal

setUnwatchMoved :: Toggle -> Config -> Config

If an object moves out of the directory being watched then stop watching it.

default: On

Internal

setOneShot :: Toggle -> Config -> Config

Watch the object only for one event and then remove it from the watch.

default: Off

Internal

setOnlyDir :: Toggle -> Config -> Config

Watch the object only if it is a directory. This provides a race-free way to ensure that the watched object is a directory.

default: Off

Internal

data WhenExists

What to do if a watch already exists when openWatch or addToWatch is called for a path.

Internal

Constructors

AddIfExists
Do not set an existing setting to Off only set to On

ReplaceIfExists
Replace the existing settings with new settings

FailIfExists
Fail the API

setWhenExists :: WhenExists -> Config -> Config

When adding a new path to the watch, specify what to do if a watch already exists on that path.

default: FailIfExists

Internal

Events of Interest

Root Level Events

setRootDeleted :: Toggle -> Config -> Config

Report when the watched path itself gets deleted.

default: On

Internal

setRootMoved :: Toggle -> Config -> Config

Report when the watched root path itself gets renamed.

default: On

Internal

Item Level Metadata change

setMetadataChanged :: Toggle -> Config -> Config

Report when the metadata e.g. owner, permission modes, modifications times of an object changes.

default: On

Internal

Item Level Access

setAccessed :: Toggle -> Config -> Config

Report when a file is accessed.

default: On

Internal

setOpened :: Toggle -> Config -> Config

Report when a file is opened.

default: On

Internal

setWriteClosed :: Toggle -> Config -> Config

Report when a file that was opened for writes is closed.

default: On

Internal

setNonWriteClosed :: Toggle -> Config -> Config

Report when a file that was opened for not writing is closed.

default: On

Internal

Item CRUD events

setCreated :: Toggle -> Config -> Config

Report when a file is created.

default: On

Internal

setDeleted :: Toggle -> Config -> Config

Report when a file is deleted.

default: On

Internal

setMovedFrom :: Toggle -> Config -> Config

Report the source of a move.

default: On

Internal

setMovedTo :: Toggle -> Config -> Config

Report the target of a move.

default: On

Internal

setModified :: Toggle -> Config -> Config

Report when a file is modified.

default: On

Internal

setAllEvents :: Toggle -> Config -> Config

Set all events On or Off.

default: On

Internal

Watch APIs

watchPathsWith :: (Config -> Config) -> NonEmpty (Array Word8) -> SerialT IO Event

Start monitoring a list of file system paths for file system events with the supplied configuration operation over the defaultConfig. The paths could be files or directories. When the path is a directory, only the files and directories directly under the watched directory are monitored, contents of subdirectories are not monitored. Monitoring starts from the current time onwards. The paths are specified as UTF-8 encoded Array of Word8.

watchPathsWith
 (setFollowSymLinks On . setUnwatchMoved Off)
 [Array.fromCString# "dir"#]

Internal

watchPaths :: NonEmpty (Array Word8) -> SerialT IO Event

Like watchPathsWith but uses the defaultConfig options.

watchPaths = watchPathsWith id

Internal

addToWatch :: Config -> Watch -> Array Word8 -> IO Watch

Given a Config and a path supplied as a UTF-8 encoded byte array, add the path to the list of paths being monitored by the given watch Handle. Returns a "watch descriptor" which is used as an idenitfier in the events generated for this particular path and also to remove the path from this watch.

Internal

removeFromWatch :: Watch -> Array Word8 -> IO Watch

Remove a path from a Watch, if a path was moved after adding you need to need to provide the original path which was used to add the Watch.

Internal

Handling Events

data Event

An Event generated by the file system. Use the accessor functions to examine the event.

Internal

getRoot :: Event -> Array Word8

Get the watch descriptor corresponding to the Event. A watch descriptor identifies which of the paths being monitored has generated this event.

Note that if a path was moved after adding to the watch, this will give the original path and not the new path after moving.

Internal

getRelPath :: Event -> Array Word8

Get the file system object path for which the event is generated, relative to the watched root. The path is a UTF-8 encoded array of bytes.

Internal

getCookie :: Event -> Cookie

Cookie is set when a rename occurs. The cookie value can be used to connect the isMovedFrom and isMovedTo events, if both the events belong to the same move operation then they will have the same cookie value.

Internal

Exception Conditions

isOverflow :: Event -> Bool

Event queue overflowed (WD is invalid for this event) and we may have lost some events.. The user application must scan everything under the watched paths to know the current state.

Internal

Root Level Events

isRootUnwatched :: Event -> Bool

A path was removed from the watch explicitly using removeFromWatch or automatically (file was deleted, or filesystem was unmounted).

Occurs only for a watched path

Internal

isRootDeleted :: Event -> Bool

Watched file/directory was itself deleted. (This event also occurs if an object is moved to another filesystem, since mv(1) in effect copies the file to the other filesystem and then deletes it from the original filesystem.) In addition, an isRootUnwatched event will subsequently be generated for the watch descriptor.

Occurs only for a watched path

Internal

isRootMoved :: Event -> Bool

Watched file/directory was itself moved within the file system.

Occurs only for a watched path

Internal

isRootUnmounted :: Event -> Bool

Filesystem containing watched object was unmounted. In addition, an isRootUnwatched event will subsequently be generated for the watch descriptor.

Occurs only for a watched path

Internal

Item Level Metadata change

isMetadataChanged :: Event -> Bool

Determine whether the event indicates inode metadata change for an object contained within the monitored path.

Metadata change may include, permissions (e.g., chmod(2)), timestamps (e.g., utimensat(2)), extended attributes (setxattr(2)), link count (since Linux 2.6.25; e.g., for the target of link(2) and for unlink(2)), and user/group ID (e.g., chown(2)).

Can occur for watched path or a file inside it

Internal

Item Level Access

isAccessed :: Event -> Bool

File was accessed (e.g. read, execve).

Occurs only for a file inside the watched directory

Internal

isOpened :: Event -> Bool

File or directory was opened.

Occurs only for a file inside the watched directory

Internal

isWriteClosed :: Event -> Bool

File opened for writing was closed.

Occurs only for a file inside the watched directory

Internal

isNonWriteClosed :: Event -> Bool

File or directory opened for read but not write was closed.

Can occur for watched path or a file inside it

Internal

Item CRUD events

isCreated :: Event -> Bool

File/directory created in watched directory (e.g., open(2) O_CREAT, mkdir(2), link(2), symlink(2), bind(2) on a UNIX domain socket).

Occurs only for an object inside the watched directory

Internal

isDeleted :: Event -> Bool

File/directory deleted from watched directory.

Occurs only for an object inside the watched directory

Internal

isMovedFrom :: Event -> Bool

Generated for the original path when an object is moved from under a monitored directory.

Occurs only for an object inside the watched directory

Internal

isMovedTo :: Event -> Bool

Generated for the new path when an object is moved under a monitored directory.

Occurs only for an object inside the watched directory

Internal

isModified :: Event -> Bool

Determine whether the event indicates modification of an object within the monitored path. This event is generated only for files and not directories.

Occurs only for an object inside the watched directory

Internal

Item Path info

isDir :: Event -> Bool

Determine whether the event is for a directory path.

Internal

Debugging

showEvent :: Event -> String

Convert an Event record to a String representation.