Streamly.Internal.FileSystem.FDIO

Description

Low level IO routines interfacing the operating system.

Documentation

write :: FD -> Ptr Word8 -> Int -> CSize -> IO CInt

write FD buffer offset length tries to write data on the given filesystem fd (cannot be a socket) up to sepcified length starting from the given offset in the buffer. The write will not block the OS thread, it may suspend the Haskell thread until write can proceed. Returns the actual amount of data written.

writeAll :: FD -> Ptr Word8 -> Int -> IO ()

Keep writing in a loop until all data in the buffer has been written.

data IOVec

Constructors

IOVec

Fields

iovBase :: !(Ptr Word8)

iovLen :: !Word64

writev :: FD -> Ptr IOVec -> Int -> IO CInt

write FD iovec count tries to write data on the given filesystem fd (cannot be a socket) from an iovec with specified number of entries. The write will not block the OS thread, it may suspend the Haskell thread until write can proceed. Returns the actual amount of data written.

writevAll :: FD -> Ptr IOVec -> Int -> IO ()

Keep writing an iovec in a loop until all the iovec entries are written.