Streamly.Internal.Mutable.Prim.Var

Description

A mutable variable in a mutation capable monad (IO/ST) holding a Prim value. This allows fast modification because of unboxed storage.

Multithread Consistency Notes

In general, any value that straddles a machine word cannot be guaranteed to be consistently read from another thread without a lock. GHC heap objects are always machine word aligned, therefore, a Var is also word aligned. On a 64-bit platform, writing a 64-bit aligned type from one thread and reading it from another thread should give consistent old or new value. The same holds true for 32-bit values on a 32-bit platform.

Documentation

data Var m a

A Var holds a single Prim value.

type MonadMut = PrimMonad

A monad that allows mutable operations using a state token.

class Prim a

Minimal complete definition

sizeOf#, alignment#, indexByteArray#, readByteArray#, writeByteArray#, setByteArray#, indexOffAddr#, readOffAddr#, writeOffAddr#, setOffAddr#

Construction

newVar :: forall m a. (MonadMut m, Prim a) => a -> m (Var m a)

Create a new mutable variable.

Write

writeVar :: (MonadMut m, Prim a) => Var m a -> a -> m ()

Write a value to a mutable variable.

modifyVar' :: (MonadMut m, Prim a) => Var m a -> (a -> a) -> m ()

Modify the value of a mutable variable using a function with strict application.

Read

readVar :: (MonadMut m, Prim a) => Var m a -> m a

Read a value from a variable.