Streamly.Internal.Network.Socket

Description

Documentation

data SockSpec

Specify the socket protocol details.

Constructors

SockSpec

Fields

sockFamily :: !Family

sockType :: !SocketType

sockProto :: !ProtocolNumber

sockOpts :: ![(SocketOption, Int)]

Use a socket

forSocketM :: (MonadMask m, MonadIO m) => (Socket -> m ()) -> Socket -> m ()

forSocketM action socket runs the monadic computation action passing the socket handle to it. The handle will be closed on exit from forSocketM, whether by normal termination or by raising an exception. If closing the handle raises an exception, then this exception will be raised by forSocketM rather than any exception raised by action.

Since: 0.8.0

withSocket :: (IsStream t, MonadAsync m, MonadCatch m) => Socket -> (Socket -> t m a) -> t m a

Like forSocketM but runs a streaming computation instead of a monadic computation.

Inhibits stream fusion

Internal

Accept connections

accept :: MonadIO m => Unfold m (Int, SockSpec, SockAddr) Socket

Unfold a three tuple (listenQLen, spec, addr) into a stream of connected protocol sockets corresponding to incoming connections. listenQLen is the maximum number of pending connections in the backlog. spec is the socket protocol and options specification and addr is the protocol address where the server listens for incoming connections.

Since: 0.7.0

connections :: MonadAsync m => Int -> SockSpec -> SockAddr -> SerialT m Socket

Start a TCP stream server that listens for connections on the supplied server address specification (address family, local interface IP address and port). The server generates a stream of connected sockets. The first argument is the maximum number of pending connections in the backlog.

Pre-release

connect :: SockSpec -> SockAddr -> IO Socket

Connect to a remote host using the given socket specification and remote address. Returns a connected socket or throws an exception.

Pre-release

connectFrom :: SockSpec -> SockAddr -> SockAddr -> IO Socket

Connect to a remote host using the given socket specification, a local address to bind to and a remote address to connect to. Returns a connected socket or throws an exception.

Pre-release

Read from connection

read :: MonadIO m => Unfold m Socket Word8

Unfolds a Socket into a byte stream. IO requests to the socket are performed in sizes of defaultChunkSize.

Since: 0.7.0

readWithBufferOf :: MonadIO m => Unfold m (Int, Socket) Word8

Unfolds the tuple (bufsize, socket) into a byte stream, read requests to the socket are performed using buffers of bufsize.

Since: 0.7.0

readChunk :: Int -> Socket -> IO (Array Word8)

Read a byte array from a file handle up to a maximum of the requested size. If no data is available on the handle it blocks until some data becomes available. If data is available then it immediately returns that data without blocking.

Since: 0.8.0

readChunks :: MonadIO m => Unfold m Socket (Array Word8)

Unfolds a socket into a stream of Word8 arrays. Requests to the socket are performed using a buffer of size defaultChunkSize. The size of arrays in the resulting stream are therefore less than or equal to defaultChunkSize.

Since: 0.7.0

readChunksWithBufferOf :: MonadIO m => Unfold m (Int, Socket) (Array Word8)

Unfold the tuple (bufsize, socket) into a stream of Word8 arrays. Read requests to the socket are performed using a buffer of size bufsize. The size of an array in the resulting stream is always less than or equal to bufsize.

Since: 0.7.0

toChunksWithBufferOf :: (IsStream t, MonadIO m) => Int -> Socket -> t m (Array Word8)

toChunksWithBufferOf size h reads a stream of arrays from file handle h. The maximum size of a single array is limited to size. fromHandleArraysUpto ignores the prevailing TextEncoding and NewlineMode on the Handle.

toChunks :: (IsStream t, MonadIO m) => Socket -> t m (Array Word8)

toChunks h reads a stream of arrays from socket handle h. The maximum size of a single array is limited to defaultChunkSize.

Since: 0.7.0

toBytes :: (IsStream t, MonadIO m) => Socket -> t m Word8

Generate a stream of elements of the given type from a socket. The stream ends when EOF is encountered.

Since: 0.7.0

Write to connection

write :: MonadIO m => Socket -> Fold m Word8 ()

Write a byte stream to a socket. Accumulates the input in chunks of up to defaultChunkSize bytes before writing.

write = writeWithBufferOf defaultChunkSize

Since: 0.7.0

writeWithBufferOf :: MonadIO m => Int -> Socket -> Fold m Word8 ()

Write a byte stream to a socket. Accumulates the input in chunks of specified number of bytes before writing.

Since: 0.7.0

writeMaybesWithBufferOf :: MonadIO m => Int -> Socket -> Fold m (Maybe Word8) ()

Write a stream of Maybe values. Keep buffering the Just values in an array. Write the array to the Handle as soon as a Nothing is encountered or the buffer size exceeds the specified limit.

Pre-release

putChunks :: (MonadIO m, Storable a) => Socket -> SerialT m (Array a) -> m ()

Write a stream of arrays to a handle.

Since: 0.7.0

putBytesWithBufferOf :: MonadIO m => Int -> Socket -> SerialT m Word8 -> m ()

Like write but provides control over the write buffer. Output will be written to the IO device as soon as we collect the specified number of input elements.

Since: 0.7.0

putBytes :: MonadIO m => Socket -> SerialT m Word8 -> m ()

Write a byte stream to a file handle. Combines the bytes in chunks of size up to defaultChunkSize before writing. Note that the write behavior depends on the IOMode and the current seek position of the handle.

Since: 0.7.0

writeChunk :: Storable a => Socket -> Array a -> IO ()

Write an Array to a file handle.

Since: 0.8.0

writeChunks :: (MonadIO m, Storable a) => Socket -> Fold m (Array a) ()

Write a stream of arrays to a socket. Each array in the stream is written to the socket as a separate IO request.

Since: 0.7.0

writeChunksWithBufferOf :: (MonadIO m, Storable a) => Int -> Socket -> Fold m (Array a) ()

writeChunksWithBufferOf bufsize socket writes a stream of arrays to socket after coalescing the adjacent arrays in chunks of bufsize. Multiple arrays are coalesed as long as the total size remains below the specified size. It never splits an array, if a single array is bigger than the specified size it emitted as it is.

Since: 0.8.0