Streamly.Internal.System.IO

Description

Documentation

defaultChunkSize :: Int

Default maximum buffer size in bytes, for reading from and writing to IO devices, the value is 32KB minus GHC allocation overhead, which is a few bytes, so that the actual allocation is 32KB.

arrayPayloadSize :: Int -> Int

When we allocate a byte array of size k the allocator actually allocates memory of size k + byteArrayOverhead. arrayPayloadSize n returns the size of the array in bytes that would result in an allocation of n bytes.

unsafeInlineIO :: IO a -> a