Streamly.Internal.Data.Fold.Tee

Description

A newtype wrapper over the Fold type providing distributing Applicative, Semigroup, Monoid, Num, Floating and Fractional instances.

Documentation

newtype Tee m a b

Tee is a newtype wrapper over the Fold type providing distributing Applicative, Semigroup, Monoid, Num, Floating and Fractional instances.

Since: 0.8.0

Constructors

Tee

Fields

toFold :: Fold m a b