Streamly.Internal.Data.Parser

Description

Fast backtracking parsers with stream fusion and native streaming capability.

Applicative and Alternative type class based combinators from the parser-combinators package can also be used with the Parser type. However, there are two important differences between parser-combinators and the equivalent ones provided in this module in terms of performance:

1) parser-combinators use plain Haskell lists to collect the results, in a strict Monad like IO, the results are necessarily buffered before they can be consumed. This may not perform optimally in streaming applications processing large amounts of data. Equivalent combinators in this module can consume the results of parsing using a Fold, thus providing a scalability and a composable consumer.

2) Several combinators in this module can be many times faster because of stream fusion. For example, many combinator in this module is much faster than the many combinator of Alternative type class.

Errors

Failing parsers in this module throw the ParseError exception.

Naming

As far as possible, try that the names of the combinators in this module are consistent with:

base/Text.ParserCombinators.ReadP

parser-combinators

megaparsec

attoparsec

parsec

Documentation

newtype Parser m a b

A continuation passing style parser representation.

Constructors

MkParser

Fields

runParser :: forall r. Int -> (Int, Int) -> ((Int, Int) -> Parse b -> m (Driver m a r)) -> m (Driver m a r)

newtype ParseError

This exception is used for two purposes:

When a parser ultimately fails, the user of the parser is intimated via this exception.

When the "extract" function of a parser needs to throw an error.

Internal

Constructors

ParseError String

data Step s b

The return type of a Parser step.

The parse operation feeds the input stream to the parser one element at a time, representing a parse Step. The parser may or may not consume the item and returns a result. If the result is Partial we can either extract the result or feed more input to the parser. If the result is Continue, we must feed more input in order to get a result. If the parser returns Done then the parser can no longer take any more input.

If the result is Continue, the parse operation retains the input in a backtracking buffer, in case the parser may ask to backtrack in future. Whenever a 'Partial n' result is returned we first backtrack by n elements in the input and then release any remaining backtracking buffer. Similarly, 'Continue n' backtracks to n elements before the current position and starts feeding the input from that point for future invocations of the parser.

If parser is not yet done, we can use the extract operation on the state of the parser to extract a result. If the parser has not yet yielded a result, the operation fails with a ParseError exception. If the parser yielded a Partial result in the past the last partial result is returned. Therefore, if a parser yields a partial result once it cannot fail later on.

The parser can never backtrack beyond the position where the last partial result left it at. The parser must ensure that the backtrack position is always after that.

Internal

Constructors

Partial Int s
Partial result with an optional backtrack request.

Partial count state means a partial result is available which can be extracted successfully, state is the opaque state of the parser to be supplied to the next invocation of the step operation. The current input position is reset to count elements back and any input before that is dropped from the backtrack buffer.

Continue Int s
Need more input with an optional backtrack request.

Continue count state means the parser has consumed the current input but no new result is generated, state is the next state of the parser. The current input is retained in the backtrack buffer and the input position is reset to count elements back.

Done Int b
Done with leftover input count and result.

Done count result means the parser has finished, it will accept no more input, last count elements from the input are unused and the result of the parser is in result.

Error String
Parser failed without generating any output.

The parsing operation may backtrack to the beginning and try another alternative.

Accumulators

fromFold :: MonadCatch m => Fold m a b -> Parser m a b

Make a Parser from a Fold.

Internal

any :: MonadCatch m => (a -> Bool) -> Parser m a Bool

>>> S.parse (PR.any (== 0)) $ S.fromList [1,0,1]
> True

all :: MonadCatch m => (a -> Bool) -> Parser m a Bool

>>> S.parse (PR.all (== 0)) $ S.fromList [1,0,1]
> False

yield :: MonadCatch m => b -> Parser m a b

A parser that always yields a pure value without consuming any input.

Internal

yieldM :: MonadCatch m => m b -> Parser m a b

A parser that always yields the result of an effectful action without consuming any input.

Internal

die :: MonadCatch m => String -> Parser m a b

A parser that always fails with an error message without consuming any input.

Internal

dieM :: MonadCatch m => m String -> Parser m a b

A parser that always fails with an effectful error message and without consuming any input.

Internal

Element parsers

peek :: MonadCatch m => Parser m a a

Peek the head element of a stream, without consuming it. Fails if it encounters end of input.

>>> S.parse ((,) <$> PR.peek <*> PR.satisfy (> 0)) $ S.fromList [1]
(1,1)

peek = lookAhead (satisfy True)

Internal

eof :: MonadCatch m => Parser m a ()

Succeeds if we are at the end of input, fails otherwise.

>>> S.parse ((,) <$> PR.satisfy (> 0) <*> PR.eof) $ S.fromList [1]
> (1,())

Internal

satisfy :: MonadCatch m => (a -> Bool) -> Parser m a a

Returns the next element if it passes the predicate, fails otherwise.

>>> S.parse (PR.satisfy (== 1)) $ S.fromList [1,0,1]
> 1

Internal

maybe :: MonadCatch m => (a -> Maybe b) -> Parser m a b

Map a Maybe returning function on the next element in the stream. The parser fails if the function returns Nothing otherwise returns the Just value.

Internal

either :: MonadCatch m => (a -> Either String b) -> Parser m a b

Map an Either returning function on the next element in the stream. If the function returns 'Left err', the parser fails with the error message err otherwise returns the Right value.

Internal

Sequence parsers

Parsers chained in series, if one parser terminates the composition terminates.

Grab a sequence of input elements without inspecting them

takeBetween :: Int -> Int -> Fold m a b -> Parser m a b

takeBetween m n takes a minimum of m and a maximum of n input elements and folds them using the supplied fold.

Stops after n elements. Fails if the stream ends before m elements could be taken.

takeBetween is the most general take operation, other take operations can be defined in terms of takeBetween. For example:

take = takeBetween 0 n  -- equivalent of takeLE
take1 = takeBetween 1 n -- equivalent of takeLE1
takeEQ = takeBetween n n
takeGE = takeBetween n maxBound

Unimplemented

take :: MonadCatch m => Int -> Fold m a b -> Parser m a b

Take at most n input elements and fold them using the supplied fold.

Stops after n elements. Never fails.

>>> S.parse (PR.take 1 FL.toList) $ S.fromList [1]
[1]

>>> S.parse (PR.take (-1) FL.toList) $ S.fromList [1]
[]

S.chunksOf n f = S.parseMany (FL.take n f)

Internal

Note: this is called takeP in some parser libraries.

TODO Once we have terminating folds, this Parse should get replaced by Fold. Alternatively, we can name it "chunkOf" and the corresponding time domain combinator as "intervalOf" or even "chunk" and "interval".

takeEQ :: MonadCatch m => Int -> Fold m a b -> Parser m a b

Stops after taking exactly n input elements.

Stops - after n elements.

Fails - if the stream ends before it can collect n elements.

>>> S.parse (PR.takeEQ 4 FL.toList) $ S.fromList [1,0,1]
> "takeEQ: Expecting exactly 4 elements, got 3"

Internal

takeGE :: MonadCatch m => Int -> Fold m a b -> Parser m a b

Take at least n input elements, but can collect more.

Stops - never.

Fails - if the stream end before producing n elements.

>>> S.parse (PR.takeGE 4 FL.toList) $ S.fromList [1,0,1]
> "takeGE: Expecting at least 4 elements, got only 3"

>>> S.parse (PR.takeGE 4 FL.toList) $ S.fromList [1,0,1,0,1]
> [1,0,1,0,1]

Internal

lookAhead :: MonadCatch m => Parser m a b -> Parser m a b

Run a parser without consuming the input.

Internal

takeWhileP :: (a -> Bool) -> Parser m a b -> Parser m a b

Like takeWhile but uses a Parser instead of a Fold to collect the input. The combinator stops when the condition fails or if the collecting parser stops.

This is a generalized version of takeWhile, for example takeWhile1 can be implemented in terms of this:

takeWhile1 cond p = takeWhile cond (takeBetween 1 maxBound p)

Stops: when the condition fails or the collecting parser stops. Fails: when the collecting parser fails.

Unimplemented

takeWhile :: MonadCatch m => (a -> Bool) -> Fold m a b -> Parser m a b

Collect stream elements until an element fails the predicate. The element on which the predicate fails is returned back to the input stream.

Stops - when the predicate fails.

Fails - never.

>>> S.parse (PR.takeWhile (== 0) FL.toList) $ S.fromList [0,0,1,0,1]
> [0,0]

We can implement a breakOn using takeWhile:

breakOn p = takeWhile (not p)

Internal

Note: This is called takeWhileP and munch in some parser libraries.

takeWhile1 :: MonadCatch m => (a -> Bool) -> Fold m a b -> Parser m a b

Like takeWhile but takes at least one element otherwise fails.

Internal

sliceSepByP :: (a -> Bool) -> Parser m a b -> Parser m a b

Like sliceSepBy but uses a Parser instead of a Fold to collect the input. sliceSepByP cond parser parses a slice of the input using parser until cond succeeds or the parser stops.

This is a generalized slicing parser which can be used to implement other parsers e.g.:

sliceSepByMax cond n p = sliceBy cond (take n p)
sliceSepByBetween cond m n p = sliceBy cond (takeBetween m n p)

Unimplemented

sliceSepBy :: MonadCatch m => (a -> Bool) -> Fold m a b -> Parser m a b

Split on an infixed separator element, dropping the separator. Splits the stream on separator elements determined by the supplied predicate, separator is considered as infixed between two segments, if one side of the separator is missing then it is parsed as an empty stream. The supplied Fold is applied on the split segments. With - representing non-separator elements and . as separator, splitOn splits as follows:

"--.--" => "--" "--"
"--."   => "--" ""
".--"   => ""   "--"

PR.sliceSepBy (== x) is an inverse of S.intercalate (S.yield x)

Let's use the following definition for illustration:

splitOn p = PR.many FL.toList $ PR.sliceSepBy p (FL.toList)
splitOn' p = S.parse (splitOn p) . S.fromList

>>> splitOn' (== '.') ""
[""]

>>> splitOn' (== '.') "."
["",""]

>>> splitOn' (== '.') ".a"
> ["","a"]

>>> splitOn' (== '.') "a."
> ["a",""]

>>> splitOn' (== '.') "a.b"
> ["a","b"]

>>> splitOn' (== '.') "a..b"
> ["a","","b"]

Stops - when the predicate succeeds.

Fails - never.

Internal

sliceSepByMax :: MonadCatch m => (a -> Bool) -> Int -> Fold m a b -> Parser m a b

Like sliceSepBy but terminates a parse even before the separator is encountered if its size exceeds the specified maximum limit.

take n = PR.sliceSepByMax (const True) n
sliceSepBy p = PR.sliceSepByMax p maxBound

Let's use the following definitions for illustration:

splitOn p n = PR.many FL.toList $ PR.sliceSepByMax p n (FL.toList)
splitOn' p n = S.parse (splitOn p n) . S.fromList

>>> splitOn' (== '.') 0 ""
[""]

>>> splitOn' (== '.') 0 "a"
infinite list of empty strings

>>> splitOn' (== '.') 3 "hello.world"
["hel","lo","wor","ld"]

If the separator is found and the limit is reached at the same time then it behaves just like sliceSepBy i.e. the separator is dropped.

>>> splitOn' (== '.') 0 "."
["",""]

>>> splitOn' (== '.') 0 ".."
["","",""]

Stops - when the predicate succeeds or the limit is reached.

Fails - never.

Internal

sliceEndWith :: (a -> Bool) -> Fold m a b -> Parser m a b

Collect stream elements until an element succeeds the predicate. Also take the element on which the predicate succeeded. The succeeding element is treated as a suffix separator which is kept in the output segement.

Stops - when the predicate succeeds.

Fails - never.

S.splitWithSuffix pred f = S.parseMany (PR.sliceEndWith pred f)

Unimplemented

sliceBeginWith :: (a -> Bool) -> Fold m a b -> Parser m a b

Collect stream elements until an elements passes the predicate, return the last element on which the predicate succeeded back to the input stream. If the predicate succeeds on the first element itself then it is kept in the stream and we continue collecting. The succeeding element is treated as a prefix separator which is kept in the output segement.

Stops - when the predicate succeeds in non-leading position.

Fails - never.

S.splitWithPrefix pred f = S.parseMany (PR.sliceBeginWith pred f)

Unimplemented

sliceSepWith :: (a -> Bool) -> Fold m a b -> Parser m a b

Like sliceSepBy but does not drop the separator element, instead separator is emitted as a separate element in the output.

Unimplemented

escapedSliceSepBy :: (a -> Bool) -> (a -> Bool) -> Fold m a b -> Parser m a b

Like sliceSepBy but the separator elements can be escaped using an escape char determined by the second predicate.

Unimplemented

escapedFrameBy :: (a -> Bool) -> (a -> Bool) -> (a -> Bool) -> Fold m a b -> Parser m a b

escapedFrameBy begin end escape parses a string framed using begin and end as the frame begin and end marker elements and escape as an escaping element to escape the occurrence of the framing elements within the frame. Nested frames are allowed, but nesting is removed when parsing.

For example,

>>> escapedFrameBy (== '{') (== '}') (== '\\') S.toList $ S.fromList "{hello}"
> "hello"

>>> escapedFrameBy (== '{') (== '}') (== '\\') S.toList $ S.fromList "{hello {world}}"
> "hello world"

>>> escapedFrameBy (== '{') (== '}') (== '\\') S.toList $ S.fromList "{hello \\{world\\}}"
> "hello {world}"

>>> escapedFrameBy (== '{') (== '}') (== '\\') S.toList $ S.fromList "{hello {world}"
> ParseError "Unterminated '{'"

Unimplemented

wordBy :: (a -> Bool) -> Fold m a b -> Parser m a b

Like splitOn but strips leading, trailing, and repeated separators. Therefore, ".a..b." having . as the separator would be parsed as ["a","b"]. In other words, its like parsing words from whitespace separated text.

Stops - when it finds a word separator after a non-word element

Fails - never.

S.wordsBy pred f = S.parseMany (PR.wordBy pred f)

Unimplemented

groupBy :: (a -> a -> Bool) -> Fold m a b -> Parser m a b

groupBy cmp f $ S.fromList [a,b,c,...] assigns the element a to the first group, then if a `cmp` b is True b is also assigned to the same group. If a `cmp` c is True then c is also assigned to the same group and so on. When the comparison fails a new group is started. Each group is folded using the Fold f and the result of the fold is emitted in the output stream.

Stops - when the comparison fails.

Fails - never.

S.groupsBy cmp f = S.parseMany (PR.groupBy cmp f)

Unimplemented

eqBy :: MonadCatch m => (a -> a -> Bool) -> [a] -> Parser m a ()

Match the given sequence of elements using the given comparison function.

>>> S.parse $ S.eqBy (==) "string" $ S.fromList "string"

>>> S.parse $ S.eqBy (==) "mismatch" $ S.fromList "match"
> *** Exception: ParseError "eqBy: failed, yet to match 7 elements"

Internal

Unimplemented

, prefixOf -- match any prefix of a given string
, suffixOf -- match any suffix of a given string
, infixOf -- match any substring of a given string

Binary Combinators

Sequential Applicative

splitWith :: MonadCatch m => (a -> b -> c) -> Parser m x a -> Parser m x b -> Parser m x c

Sequential parser application. Apply two parsers sequentially to an input stream. The input is provided to the first parser, when it is done the remaining input is provided to the second parser. If both the parsers succeed their outputs are combined using the supplied function. The operation fails if any of the parsers fail.

Note: This is a parsing dual of appending streams using serial, it splits the streams using two parsers and zips the results.

This implementation is strict in the second argument, therefore, the following will fail:

>>> S.parse (PR.splitWith const (PR.satisfy (> 0)) undefined) $ S.fromList [1]

Compare with Applicative instance method <*>. This implementation allows stream fusion but has quadratic complexity. This can fuse with other operations and can be faster than Applicative instance for small number (less than 8) of compositions.

Internal

split_ :: MonadCatch m => Parser m x a -> Parser m x b -> Parser m x b

Sequential parser application ignoring the output of the first parser. Apply two parsers sequentially to an input stream. The input is provided to the first parser, when it is done the remaining input is provided to the second parser. The output of the parser is the output of the second parser. The operation fails if any of the parsers fail.

This implementation is strict in the second argument, therefore, the following will fail:

>>> S.parse (split_ (PR.satisfy (> 0)) undefined) $ S.fromList [1]

Compare with Applicative instance method *>. This implementation allows stream fusion but has quadratic complexity. This can fuse with other operations, and can be faster than Applicative instance for small number (less than 8) of compositions.

Internal

Parallel Applicatives

teeWith :: MonadCatch m => (a -> b -> c) -> Parser m x a -> Parser m x b -> Parser m x c

teeWith f p1 p2 distributes its input to both p1 and p2 until both of them succeed or anyone of them fails and combines their output using f. The parser succeeds if both the parsers succeed.

Internal

teeWithFst :: MonadCatch m => (a -> b -> c) -> Parser m x a -> Parser m x b -> Parser m x c

Like teeWith but ends parsing and zips the results, if available, whenever the first parser ends.

Internal

teeWithMin :: MonadCatch m => (a -> b -> c) -> Parser m x a -> Parser m x b -> Parser m x c

Like teeWith but ends parsing and zips the results, if available, whenever any of the parsers ends or fails.

Unimplemented

Sequential Interleaving

deintercalate :: Fold m a y -> Parser m x a -> Fold m b z -> Parser m x b -> Parser m x (y, z)

Apply two parsers alternately to an input stream. The input stream is considered an interleaving of two patterns. The two parsers represent the two patterns.

This undoes a "gintercalate" of two streams.

Unimplemented

Sequential Alternative

alt :: MonadCatch m => Parser m x a -> Parser m x a -> Parser m x a

Sequential alternative. Apply the input to the first parser and return the result if the parser succeeds. If the first parser fails then backtrack and apply the same input to the second parser and return the result.

Note: This implementation is not lazy in the second argument. The following will fail:

>>> S.parse (PR.satisfy (> 0) `PR.alt` undefined) $ S.fromList [1..10]

Compare with Alternative instance method <|>. This implementation allows stream fusion but has quadratic complexity. This can fuse with other operations and can be much faster than Alternative instance for small number (less than 8) of alternatives.

Internal

Parallel Alternatives

shortest :: MonadCatch m => Parser m x a -> Parser m x a -> Parser m x a

Shortest alternative. Apply both parsers in parallel but choose the result from the one which consumed least input i.e. take the shortest succeeding parse.

Internal

longest :: MonadCatch m => Parser m x a -> Parser m x a -> Parser m x a

Longest alternative. Apply both parsers in parallel but choose the result from the one which consumed more input i.e. take the longest succeeding parse.

Internal

N-ary Combinators

Sequential Collection

sequence :: Fold m b c -> t (Parser m a b) -> Parser m a c

sequence f t collects sequential parses of parsers in the container t using the fold f. Fails if the input ends or any of the parsers fail.

This is same as sequence but more efficient.

Unimplemented

concatMap :: MonadCatch m => (b -> Parser m a c) -> Parser m a b -> Parser m a c

Map a Parser returning function on the result of a Parser.

Compare with Monad instance method >>=. This implementation allows stream fusion but has quadratic complexity. This can fuse with other operations and can be much faster than Monad instance for small number (less than 8) of compositions.

Internal

Sequential Repetition

count :: Int -> Fold m b c -> Parser m a b -> Parser m a c

count n f p collects exactly n sequential parses of parser p using the fold f. Fails if the input ends or the parser fails before n results are collected.

Unimplemented

countBetween :: Int -> Int -> Fold m b c -> Parser m a b -> Parser m a c

countBetween m n f p collects between m and n sequential parses of parser p using the fold f. Stop after collecting n results. Fails if the input ends or the parser fails before m results are collected.

Unimplemented

manyP :: Parser m b c -> Parser m a b -> Parser m a c

Like many but uses a Parser instead of a Fold to collect the results. Parsing stops or fails if the collecting parser stops or fails.

Unimplemented

many :: MonadCatch m => Fold m b c -> Parser m a b -> Parser m a c

Collect zero or more parses. Apply the parser repeatedly on the input stream, stop when the parser fails, accumulate zero or more parse results using the supplied Fold. This parser never fails, in case the first application of parser fails it returns an empty result.

Compare with many.

Internal

some :: MonadCatch m => Fold m b c -> Parser m a b -> Parser m a c

Collect one or more parses. Apply the supplied parser repeatedly on the input stream and accumulate the parse results as long as the parser succeeds, stop when it fails. This parser fails if not even one result is collected.

Compare with some.

Internal

manyTillP :: Parser m b c -> Parser m a b -> Parser m a x -> Parser m a c

Like manyTill but uses a Parser to collect the results instead of a Fold. Parsing stops or fails if the collecting parser stops or fails.

We can implemnent parsers like the following using manyTillP:

countBetweenTill m n f p = manyTillP (takeBetween m n f) p

Unimplemented

manyTill :: MonadCatch m => Fold m b c -> Parser m a b -> Parser m a x -> Parser m a c

manyTill f collect test tries the parser test on the input, if test fails it backtracks and tries collect, after collect succeeds test is tried again and so on. The parser stops when test succeeds. The output of test is discarded and the output of collect is accumulated by the supplied fold. The parser fails if collect fails.

Internal

manyThen :: Fold m b c -> Parser m a b -> Parser m a x -> Parser m a c

manyThen f collect recover repeats the parser collect on the input and collects the output in the supplied fold. If the the parser collect fails, parser recover is run until it stops and then we start repeating the parser collect again. The parser fails if the recovery parser fails.

For example, this can be used to find a key frame in a video stream after an error.

Unimplemented

Special cases

TODO: traditional implmentations of these may be of limited use. For example, consider parsing lines separated by \r\n. The main parser will have to detect and exclude the sequence \r\n anyway so that we can apply the "sep" parser.

We can instead implement these as special cases of deintercalate.

, endBy
, sepBy
, sepEndBy
, beginBy
, sepBeginBy
, sepAroundBy

Distribution

A simple and stupid impl would be to just convert the stream to an array and give the array reference to all consumers. The array can be grown on demand by any consumer and truncated when nonbody needs it.

Distribute to collection

Distribute to repetition

Interleaved collection

Round robin

Priority based

roundRobin :: Fold m b c -> t (Parser m a b) -> Parser m a c

Apply a collection of parsers to an input stream in a round robin fashion. Each parser is applied until it stops and then we repeat starting with the the first parser again.

Unimplemented

Collection of Alternatives

Unimplemented

, shortestN
, longestN
, fastestN -- first N successful in time
, choiceN  -- first N successful in position

choice :: t (Parser m a b) -> Parser m a b

choice parsers applies the parsers in order and returns the first successful parse.

This is same as asum but more efficient.

Unimplemented

Repeated Alternatives

retryMaxTotal :: Int -> Fold m b c -> Parser m a b -> Parser m a c

Keep trying a parser up to a maximum of n failures. When the parser fails the input consumed till now is dropped and the new instance is tried on the fresh input.

Unimplemented

retryMaxSuccessive :: Int -> Fold m b c -> Parser m a b -> Parser m a c

Like retryMaxTotal but aborts after n successive failures.

Unimplemented

retry :: Parser m a b -> Parser m a b

Keep trying a parser until it succeeds. When the parser fails the input consumed till now is dropped and the new instance is tried on the fresh input.

Unimplemented