Streamly.Internal.Data.Parser

Description

Parsers.

Documentation

data Parser m a b

A parsing fold is represented as Parser step initial extract. Before we drive a parser we call the initial action to retrieve the initial state of the fold. The driver invokes step with the state returned by the previous step and the next input element. It results into a new state and a command to the driver represented by Step type. At any point of time the driver can call extract to inspect the result of the fold. It may result in an error or an output value.

Constructors

Parser (s -> a -> m (Step s)) (m s) (s -> m (Either String b))

Combinators

fromFold :: Monad m => Fold m a b -> Parser m a b

The resulting parse never terminates and never errors out.

Folds

any :: Monad m => (a -> Bool) -> Parser m a Bool

>>> S.parse (PR.any (== 0)) $ S.fromList [1,0,1]
> Right True

all :: Monad m => (a -> Bool) -> Parser m a Bool

>>> S.parse (PR.all (== 0)) $ S.fromList [1,0,1]
> Right False

sepBy :: Monad m => (a -> Bool) -> Fold m a b -> Parser m a b

Keep taking elements until the predicate succeeds. Drop the succeeding element.

Stops - when the predicate succeeds.

Fails - never.

>>> S.parse (PR.sepBy (== 1) FL.toList) $ S.fromList [0,0,1,0,1]
> Right [0,0]

>>> S.toList $ S.splitParse (PR.sepBy (== 1) FL.toList) $ S.fromList [0,0,1,0,1]
> [[0,0],[0],[]]

S.splitOn pred f = S.splitParse (PR.sepBy pred f)

Internal

sepByMax :: Monad m => (a -> Bool) -> Int -> Fold m a b -> Parser m a b

Split using a condition or a count whichever occurs first. This is a hybrid of splitOn and take. The element on which the condition succeeds is dropped.

Internal

sepWithSuffix :: (a -> Bool) -> Fold m a b -> Parser m a b

Keep taking elements until the predicate succeeds. Take the succeeding element as well.

Stops - when the predicate succeeds.

Fails - never.

S.splitWithSuffix pred f = S.splitParse (PR.sepWithSuffix pred f)

Unimplemented

wordBy :: (a -> Bool) -> Fold m a b -> Parser m a b

Like splitOn after stripping leading, trailing, and repeated separators. Therefore, ".a..b." with . as the separator would be parsed as ["a","b"]. In other words, its like parsing words from whitespace separated text.

Stops - when it finds a word separator after a non-word element

Fails - never.

S.wordsBy pred f = S.splitParse (PR.wordBy pred f)

Unimplemented

Parsers

takeWhile :: Monad m => (a -> Bool) -> Fold m a b -> Parser m a b

Take until the predicate fails. The element on which the predicate fails is returned back to the input stream.

Stops - when the predicate fails.

Fails - never.

>>> S.parse (PR.takeWhile (== 0) FL.toList) $ S.fromList [0,0,1,0,1]
> Right [0,0]

breakOn p = takeWhile (not p)

Internal

takeEQ :: Monad m => Int -> Fold m a b -> Parser m a b

Stops after taking exactly n input elements.

Stops - after n elements.

Fails - if the stream ends before it can collect n elements.

>>> S.parse (PR.takeExact 4 FL.toList) $ S.fromList [1,0,1]
> Left "takeEQ: Expecting exactly 4 elements, got 3"

Internal

takeGE :: Monad m => Int -> Fold m a b -> Parser m a b

Take at least n input elements, but can collect more.

Stops - never.

Fails - if the stream end before producing n elements.

>>> S.parse (PR.takeGE 4 FL.toList) $ S.fromList [1,0,1]
> Left "takeGE: Expecting at least 4 elements, got only 3"

>>> S.parse (PR.takeGE 4 FL.toList) $ S.fromList [1,0,1,0,1]
> Right [1,0,1,0,1]

Internal

sepWithPrefix :: (a -> Bool) -> Fold m a b -> Parser m a b

Keep taking elements until the predicate succeeds. Return the succeeding element back to the input.

Stops - when the predicate succeeds.

Fails - never.

S.splitWithPrefix pred f = S.splitParse (PR.sepWithPrefix pred f)

Unimplemented

groupBy :: (a -> a -> Bool) -> Fold m a b -> Parser m a b

group cmp f $ S.fromList [a,b,c,...] assigns the element a to the first group, if a `cmp` b is True then b is also assigned to the same group. If a `cmp` c is True then c is also assigned to the same group and so on. When the comparison fails a new group is started. Each group is folded using the fold f and the result of the fold is emitted in the output stream.

Stops - when the group ends.

Fails - never.

S.groupsBy cmp f = S.splitParse (PR.groupBy cmp f)

Unimplemented