Streamly.Internal.Data.Parser.ParserD

Description

Direct style parser implementation with stream fusion.

Documentation

data Parser m a b

A parser is a fold that can fail and is represented as Parser step initial extract. Before we drive a parser we call the initial action to retrieve the initial state of the fold. The parser driver invokes step with the state returned by the previous step and the next input element. It results into a new state and a command to the driver represented by Step type. The driver keeps invoking the step function until it stops or fails. At any point of time the driver can call extract to inspect the result of the fold. It may result in an error or an output value.

Internal

Constructors

forall s. Parser (s -> a -> m (Step s b)) (m s) (s -> m b)

newtype ParseError

This exception is used for two purposes:

When a parser ultimately fails, the user of the parser is intimated via this exception.

When the "extract" function of a parser needs to throw an error.

Internal

Constructors

ParseError String

data Step s b

The return type of a Parser step.

The parse operation feeds the input stream to the parser one element at a time, representing a parse Step. The parser may or may not consume the item and returns a result. If the result is Partial we can either extract the result or feed more input to the parser. If the result is Continue, we must feed more input in order to get a result. If the parser returns Done then the parser can no longer take any more input.

If the result is Continue, the parse operation retains the input in a backtracking buffer, in case the parser may ask to backtrack in future. Whenever a 'Partial n' result is returned we first backtrack by n elements in the input and then release any remaining backtracking buffer. Similarly, 'Continue n' backtracks to n elements before the current position and starts feeding the input from that point for future invocations of the parser.

If parser is not yet done, we can use the extract operation on the state of the parser to extract a result. If the parser has not yet yielded a result, the operation fails with a ParseError exception. If the parser yielded a Partial result in the past the last partial result is returned. Therefore, if a parser yields a partial result once it cannot fail later on.

The parser can never backtrack beyond the position where the last partial result left it at. The parser must ensure that the backtrack position is always after that.

Internal

Constructors

Partial Int s
Partial result with an optional backtrack request.

Partial count state means a partial result is available which can be extracted successfully, state is the opaque state of the parser to be supplied to the next invocation of the step operation. The current input position is reset to count elements back and any input before that is dropped from the backtrack buffer.

Continue Int s
Need more input with an optional backtrack request.

Continue count state means the parser has consumed the current input but no new result is generated, state is the next state of the parser. The current input is retained in the backtrack buffer and the input position is reset to count elements back.

Done Int b
Done with leftover input count and result.

Done count result means the parser has finished, it will accept no more input, last count elements from the input are unused and the result of the parser is in result.

Error String
Parser failed without generating any output.

The parsing operation may backtrack to the beginning and try another alternative.

Accumulators

fromFold :: Monad m => Fold m a b -> Parser m a b

See fromFold.

Internal

any :: Monad m => (a -> Bool) -> Parser m a Bool

all :: Monad m => (a -> Bool) -> Parser m a Bool

yield :: Monad m => b -> Parser m a b

See yield.

Internal

yieldM :: Monad m => m b -> Parser m a b

See yieldM.

Internal

die :: MonadThrow m => String -> Parser m a b

See die.

Internal

dieM :: MonadThrow m => m String -> Parser m a b

See dieM.

Internal

Element parsers

peek :: MonadThrow m => Parser m a a

See peek.

Internal

eof :: Monad m => Parser m a ()

See eof.

Internal

satisfy :: MonadThrow m => (a -> Bool) -> Parser m a a

See satisfy.

Internal

maybe :: MonadThrow m => (a -> Maybe b) -> Parser m a b

See maybe.

Internal

either :: MonadThrow m => (a -> Either String b) -> Parser m a b

See either.

Internal

Sequence parsers

take :: Monad m => Int -> Fold m a b -> Parser m a b

See take.

Internal

takeEQ :: MonadThrow m => Int -> Fold m a b -> Parser m a b

See takeEQ.

Internal

takeGE :: MonadThrow m => Int -> Fold m a b -> Parser m a b

See takeGE.

Internal

lookAhead :: MonadThrow m => Parser m a b -> Parser m a b

See lookahead.

Internal

takeWhile :: Monad m => (a -> Bool) -> Fold m a b -> Parser m a b

See takeWhile.

Internal

takeWhile1 :: MonadThrow m => (a -> Bool) -> Fold m a b -> Parser m a b

See takeWhile1.

Internal

sliceSepBy :: Monad m => (a -> Bool) -> Fold m a b -> Parser m a b

See sliceSepBy.

Internal

sliceSepByMax :: Monad m => (a -> Bool) -> Int -> Fold m a b -> Parser m a b

See sliceSepByMax.

Internal

sliceEndWith :: (a -> Bool) -> Fold m a b -> Parser m a b

See sliceEndWith.

Unimplemented

sliceBeginWith :: (a -> Bool) -> Fold m a b -> Parser m a b

See sliceBeginWith.

Unimplemented

wordBy :: (a -> Bool) -> Fold m a b -> Parser m a b

See wordBy.

Unimplemented

groupBy :: (a -> a -> Bool) -> Fold m a b -> Parser m a b

See groupBy.

Unimplemented

eqBy :: MonadThrow m => (a -> a -> Bool) -> [a] -> Parser m a ()

See eqBy.

Internal

Binary Combinators

Sequential Applicative

splitWith :: Monad m => (a -> b -> c) -> Parser m x a -> Parser m x b -> Parser m x c

See splitWith.

Internal

split_ :: Monad m => Parser m x a -> Parser m x b -> Parser m x b

See split_.

Internal

Parallel Applicatives

teeWith :: Monad m => (a -> b -> c) -> Parser m x a -> Parser m x b -> Parser m x c

See teeWith.

Broken

teeWithFst :: Monad m => (a -> b -> c) -> Parser m x a -> Parser m x b -> Parser m x c

See teeWithFst.

Broken

teeWithMin :: (a -> b -> c) -> Parser m x a -> Parser m x b -> Parser m x c

See teeWithMin.

Unimplemented

Sequential Interleaving

deintercalate :: Fold m a y -> Parser m x a -> Fold m b z -> Parser m x b -> Parser m x (y, z)

See deintercalate.

Unimplemented

Sequential Alternative

alt :: Monad m => Parser m x a -> Parser m x a -> Parser m x a

See alt.

Internal

Parallel Alternatives

shortest :: Monad m => Parser m x a -> Parser m x a -> Parser m x a

See shortest.

Broken

longest :: MonadCatch m => Parser m x a -> Parser m x a -> Parser m x a

See longest.

Broken

N-ary Combinators

Sequential Collection

sequence :: Fold m b c -> t (Parser m a b) -> Parser m a c

See sequence.

Unimplemented

concatMap :: Monad m => (b -> Parser m a c) -> Parser m a b -> Parser m a c

See concatMap.

Internal

Sequential Repetition

count :: Int -> Fold m b c -> Parser m a b -> Parser m a c

See count.

Unimplemented

countBetween :: Int -> Int -> Fold m b c -> Parser m a b -> Parser m a c

See countBetween.

Unimplemented

many :: MonadCatch m => Fold m b c -> Parser m a b -> Parser m a c

See many.

Internal

some :: MonadCatch m => Fold m b c -> Parser m a b -> Parser m a c

See some.

Internal

manyTill :: MonadCatch m => Fold m b c -> Parser m a b -> Parser m a x -> Parser m a c

See manyTill.

Internal

Collection of Alternatives

choice :: t (Parser m a b) -> Parser m a b

See choice.

Unimplemented