Streamly.Internal.Data.Stream.Prelude

Description

Low level functions using StreamK as the intermediate stream type. These functions are used in SerialTAsyncTAheadT/ParallelT stream modules to implement their instances..

Running Effects

drain :: Monad m => Stream m a -> m ()

Conversion operations

fromList :: Monad m => [a] -> Stream m a

fromList = foldr cons nil

Construct a stream from a list of pure values. This is more efficient than fromFoldable for serial streams.

Since: 0.4.0

toList :: Monad m => Stream m a -> m [a]

Convert a stream into a list in the underlying monad.

Since: 0.1.0

Fold operations

foldrM :: Monad m => (a -> m b -> m b) -> m b -> Stream m a -> m b

foldrMx :: Monad m => (a -> m x -> m x) -> m x -> (m x -> m b) -> Stream m a -> m b

foldr :: Monad m => (a -> b -> b) -> b -> Stream m a -> m b

foldlx' :: Monad m => (x -> a -> x) -> x -> (x -> b) -> Stream m a -> m b

Strict left fold with an extraction function. Like the standard strict left fold, but applies a user supplied extraction function (the third argument) to the folded value at the end. This is designed to work with the foldl library. The suffix x is a mnemonic for extraction.

Since: 0.7.0

foldlMx' :: Monad m => (x -> a -> m x) -> m x -> (x -> m b) -> Stream m a -> m b

Like foldlx', but with a monadic step function.

Since: 0.7.0

foldl' :: Monad m => (b -> a -> b) -> b -> Stream m a -> m b

Strict left associative fold.

Since: 0.2.0

fold :: Monad m => Fold m a b -> Stream m a -> m b

Zip style operations

eqBy :: Monad m => (a -> b -> Bool) -> Stream m a -> Stream m b -> m Bool

Compare two streams for equality

Since: 0.5.3

cmpBy :: Monad m => (a -> b -> Ordering) -> Stream m a -> Stream m b -> m Ordering

Compare two streams

Since: 0.5.3