Streamly.Internal.Data.Stream.StreamD.Type

Description

Copyright : (c) Roman Leshchinskiy 2008-2010

License : BSD3 Maintainer : streamly@composewell.com Stability : experimental Portability : GHC

The stream type

data Step s a

A stream is a succession of Steps. A Yield produces a single value and the next state of the stream. Stop indicates there are no more values in the stream.

Constructors

Yield a s

Skip s

Stop

data Stream m a

A stream consists of a step function that generates the next step given a current state, and the current state.

Constructors

UnStream (State Stream m a -> s -> m (Step s a)) s

Bundled Patterns

pattern Stream :: (State Stream m a -> s -> m (Step s a)) -> s -> Stream m a

fromStreamK :: Monad m => Stream m a -> Stream m a

toStreamK :: Monad m => Stream m a -> Stream m a

fromStreamD :: (IsStream t, Monad m) => Stream m a -> t m a

map :: Monad m => (a -> b) -> Stream m a -> Stream m b

mapM :: Monad m => (a -> m b) -> Stream m a -> Stream m b

Map a monadic function over a Stream

yield :: Monad m => a -> Stream m a

Create a singleton Stream from a pure value.

yieldM :: Monad m => m a -> Stream m a

Create a singleton Stream from a monadic action.

concatMap :: Monad m => (a -> Stream m b) -> Stream m a -> Stream m b

concatMapM :: Monad m => (a -> m (Stream m b)) -> Stream m a -> Stream m b

foldrT :: (Monad m, Monad (t m), MonadTrans t) => (a -> t m b -> t m b) -> t m b -> Stream m a -> t m b

foldrM :: Monad m => (a -> m b -> m b) -> m b -> Stream m a -> m b

foldrMx :: Monad m => (a -> m x -> m x) -> m x -> (m x -> m b) -> Stream m a -> m b

foldr :: Monad m => (a -> b -> b) -> b -> Stream m a -> m b

foldrS :: Monad m => (a -> Stream m b -> Stream m b) -> Stream m b -> Stream m a -> Stream m b

foldl' :: Monad m => (b -> a -> b) -> b -> Stream m a -> m b

foldlM' :: Monad m => (b -> a -> m b) -> b -> Stream m a -> m b

foldlx' :: Monad m => (x -> a -> x) -> x -> (x -> b) -> Stream m a -> m b

foldlMx' :: Monad m => (x -> a -> m x) -> m x -> (x -> m b) -> Stream m a -> m b

toList :: Monad m => Stream m a -> m [a]

fromList :: Monad m => [a] -> Stream m a

Convert a list of pure values to a Stream

eqBy :: Monad m => (a -> b -> Bool) -> Stream m a -> Stream m b -> m Bool

cmpBy :: Monad m => (a -> b -> Ordering) -> Stream m a -> Stream m b -> m Ordering

Compare two streams lexicographically

take :: Monad m => Int -> Stream m a -> Stream m a

data GroupState s fs

Constructors

GroupStart s

GroupBuffer s fs Int

GroupYield fs (GroupState s fs)

GroupFinish

groupsOf :: Monad m => Int -> Fold m a b -> Stream m a -> Stream m b

groupsOf2 :: Monad m => Int -> m c -> Fold2 m c a b -> Stream m a -> Stream m b