Streamly.Internal.Data.Stream.Zip

Description

Documentation

data ZipSerialM m a

The applicative instance of ZipSerialM zips a number of streams serially i.e. it produces one element from each stream serially and then zips all those elements.

main = (toList . zipSerially $ (,,) <$> s1 <*> s2 <*> s3) >>= print
    where s1 = fromFoldable [1, 2]
          s2 = fromFoldable [3, 4]
          s3 = fromFoldable [5, 6]

[(1,3,5),(2,4,6)]

The Semigroup instance of this type works the same way as that of SerialT.

Since: 0.2.0

type ZipSerial = ZipSerialM IO

An IO stream whose applicative instance zips streams serially.

Since: 0.2.0

zipSerially :: IsStream t => ZipSerialM m a -> t m a

Fix the type of a polymorphic stream as ZipSerialM.

Since: 0.2.0

data ZipAsyncM m a

Like ZipSerialM but zips in parallel, it generates all the elements to be zipped concurrently.

main = (toList . zipAsyncly $ (,,) <$> s1 <*> s2 <*> s3) >>= print
    where s1 = fromFoldable [1, 2]
          s2 = fromFoldable [3, 4]
          s3 = fromFoldable [5, 6]

[(1,3,5),(2,4,6)]

The Semigroup instance of this type works the same way as that of SerialT.

Since: 0.2.0

type ZipAsync = ZipAsyncM IO

An IO stream whose applicative instance zips streams wAsyncly.

Since: 0.2.0

zipAsyncly :: IsStream t => ZipAsyncM m a -> t m a

Fix the type of a polymorphic stream as ZipAsyncM.

Since: 0.2.0

zipWith :: (IsStream t, Monad m) => (a -> b -> c) -> t m a -> t m b -> t m c

Zip two streams serially using a pure zipping function.

> S.toList $ S.zipWith (+) (S.fromList [1,2,3]) (S.fromList [4,5,6])
[5,7,9]

Since: 0.1.0

zipWithM :: (IsStream t, Monad m) => (a -> b -> m c) -> t m a -> t m b -> t m c

Like zipWith but using a monadic zipping function.

Since: 0.4.0

zipAsyncWith :: (IsStream t, MonadAsync m) => (a -> b -> c) -> t m a -> t m b -> t m c

Like zipWith but zips concurrently i.e. both the streams being zipped are generated concurrently.

Since: 0.1.0

zipAsyncWithM :: (IsStream t, MonadAsync m) => (a -> b -> m c) -> t m a -> t m b -> t m c

Like zipWithM but zips concurrently i.e. both the streams being zipped are generated concurrently.

Since: 0.4.0

Deprecated

type ZipStream = ZipSerialM

Deprecated: Please use ZipSerialM instead.

Since: 0.1.0

zipping :: IsStream t => ZipSerialM m a -> t m a

Deprecated: Please use zipSerially instead.

Same as zipSerially.

Since: 0.1.0

zippingAsync :: IsStream t => ZipAsyncM m a -> t m a

Deprecated: Please use zipAsyncly instead.

Same as zipAsyncly.

Since: 0.1.0