Streamly.Internal.Data.Zipper

Description

Zipper for parsers.

Documentation

data Zipper m a

Constructors

Zipper [Int] [a] [a] (Stream m a)

Construction

nil :: Zipper m a

Empty zipper.

Internal

fromStream :: Stream m a -> Zipper m a

Create a zipper from a stream.

Internal

fromList :: [a] -> Zipper m a

Create a zipper from a list.

Internal

Checkpointing

checkpoint :: Zipper m a -> Zipper m a

Add a checkpoint to the Zipper so that we do not release the buffer beyond the checkpoint.

Internal

release :: Zipper m a -> Zipper m a

Release the latest checkpoint, we can now release any values held by the checkpoint.

Internal

restore :: Zipper m a -> Zipper m a

Rewind to restore the cursor to the latest checkpoint.

Internal

Parsing

parse :: MonadCatch m => (s -> a -> m (Step s b)) -> m s -> (s -> m b) -> Zipper m a -> m (Zipper m a, Either String b)